% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{abort_f}
\alias{abort_f}
\title{Semantic sugar to wrap an abort statement in a function}
\usage{
abort_f(...)
}
\description{
This can be used to provide abort statements as a direct argument to
tryCatch without wrapping them in a function call
}
