% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{add_R_variability}
\alias{add_R_variability}
\title{Add change point model for the variability of Rt}
\usage{
add_R_variability(
  length_R,
  h,
  length_seeding,
  partial_window,
  partial_generation,
  changepoint_dist,
  modeldata
)
}
\arguments{
\item{length_R}{Length of the modeled Rt time series}

\item{changepoint_dist}{How many days should the change points be apart? If
zero, no change points are modeled (fixed R variability).}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Add change point model for the variability of Rt
}
\details{
The change points are placed going backwards from the end of the
time series, i.e. on the last day, then \code{changepoint_dist} days before
that, and so on. The distance between the first and second change point at
the start of the time series (partly falls into the seeding phase) varies
between half and double the \code{changepoint_dist}.
}
\keyword{internal}
