% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{apply_component}
\alias{apply_component}
\title{Execute one component body against the modeldata environment}
\usage{
apply_component(comp, md)
}
\description{
Execute one component body against the modeldata environment
}
\keyword{internal}
