% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{as_modeldata_list}
\alias{as_modeldata_list}
\title{Convert the compile-time environment to a plain modeldata list}
\usage{
as_modeldata_list(md)
}
\description{
Convert the compile-time environment to a plain modeldata list
}
\keyword{internal}
