% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_smooth.R
\name{changepoint_dirichlet_prior}
\alias{changepoint_dirichlet_prior}
\title{Solve Dirichlet Parameters for Soft Changepoint Model}
\usage{
changepoint_dirichlet_prior(
  alpha,
  n,
  min_break_dist = 1,
  prior_previous = NULL
)
}
\arguments{
\item{alpha}{Total sum of all p_i}

\item{n}{Number of parameters to solve for}

\item{min_break_dist}{Minimum distance between changepoints. For first
segment, can be left empty/set to 1.}

\item{prior_previous}{Vector with alpha values of prior for previous segment.
Required if prior_previous>1.}
}
\value{
A vector of p_i values
}
\description{
This function solves for the parameters p_i of a Dirichlet
distribution given a total sum alpha and number of parameters n.
}
\keyword{internal}
