% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy.R
\name{check_provided}
\alias{check_provided}
\title{Check if all required data and assumptions have been provided.}
\usage{
check_provided(
  f_env = list(),
  data = list(),
  assumptions = list(),
  required_data = c(),
  required_assumptions = c()
)
}
\arguments{
\item{f_env}{The environment to be checked}

\item{data}{A list with data}

\item{assumptions}{A list with assumptions}

\item{required_data}{A character vector with the names of required data}

\item{required_assumptions}{A character vector with the names of required
assumptions}
}
\value{
TRUE if all required data and assumptions are present in \code{f_env}.
}
\description{
Check if all required data and assumptions have been provided.
}
