% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{cli_help}
\alias{cli_help}
\title{Print link to help page of function in CLI}
\usage{
cli_help(function_name, label = NULL)
}
\arguments{
\item{function_name}{Name of the function}

\item{label}{Optional label for the link, if NULL (default), the function
name is used as label}
}
\value{
A string with a link to the help page of the function
}
\description{
Print link to help page of function in CLI
}
\keyword{internal}
