% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{compareNA}
\alias{compareNA}
\title{Compare entries of two vectors with potentially missing values}
\usage{
compareNA(v1, v2)
}
\arguments{
\item{v1}{First vector}

\item{v2}{Second vector, should be of same length as first one}
}
\value{
A boolean vector with element-wise comparisons. Comparisons including
NAs are coded as FALSE.
}
\description{
Compare entries of two vectors with potentially missing values
}
