% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{component_functions_}
\alias{component_functions_}
\title{Get modeling functions for a component (internal function)}
\usage{
component_functions_(
  component,
  collapse = "\\n",
  prefix = "- [",
  suffix = "()]"
)
}
\arguments{
\item{component}{A character vector with the name of the component}

\item{collapse}{A character used to collapse helpers into one string. Default
is NULL (do not collapse).}

\item{prefix}{A prefix to add to each function name}

\item{suffix}{A suffix to add to each function name}
}
\value{
A character vector with all available modeling functions for the
component. If collapse is not NULL, the helper functions are collapsed into
a single string.
}
\description{
Get all available modeling functions for a given component. Argument defaults
are such that helpers are returned as a comma-separated list of roxygen
function links.
}
