% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{component_helpers}
\alias{component_helpers}
\title{Get model helpers for a component}
\usage{
component_helpers(component)
}
\arguments{
\item{component}{A character vector with the name of the component}
}
\value{
A character vector with all available model helper functions for
the component.
}
\description{
Get model helpers for a component
}
\examples{
component_helpers("R")
}
