% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{component_str_details}
\alias{component_str_details}
\title{Determine display details for a component (non-default scalar arguments)}
\usage{
component_str_details(helper, slot_spec)
}
\description{
Determine display details for a component (non-default scalar arguments)
}
\keyword{internal}
