% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy.R
\name{copy_env}
\alias{copy_env}
\title{Copy an environment}
\usage{
copy_env(env, exclude = c())
}
\arguments{
\item{env}{The environment to be copied}

\item{exclude}{Character vector with names to be excluded in the copy}
}
\value{
A copy of the environment
}
\description{
Copy an environment
}
\details{
The new environment has the calling environment of \code{copy_env} as its
parent. This ensure that all functions in the namespace are accessible.
}
