% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{dampening_assume}
\alias{dampening_assume}
\title{Dampen forecasts}
\usage{
dampening_assume(dampening = 0.8, modeldata = modeldata_init())
}
\arguments{
\item{dampening}{The forecast dampening parameter. A value of 1 means no
dampening, a value of 0 means flat forecast. The default is 0.8, which
levels off after a horizon of approximately 2 weeks.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option dampens the forecast of Rt so that trends in
transmission will level off after some time. This prevents unrealistic
extrapolation of transmission dynamics.
}
\details{
The applied dampening is exponential, i.e. the trend is reduced by
\code{dampening^1} on the first forecast day, by \code{dampening^2} on the second
forecast day, and so on.
}
