% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{damping_assume}
\alias{damping_assume}
\title{Dampen forecasts}
\usage{
damping_assume(damping = 0.95)
}
\arguments{
\item{damping}{The forecast damping parameter. A value of 1 means no
damping, a value of 0 means flat forecast. The default is 0.8, which
levels off after a horizon of approximately 2 weeks.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option dampens the forecast of Rt so that trends in
transmission will level off after some time. This prevents unrealistic
extrapolation of transmission dynamics.
}
\details{
The applied damping is exponential, i.e. the trend is reduced by
\code{damping^1} on the first forecast day, by \code{damping^2} on the second
forecast day, and so on.
}
