% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{damping_none}
\alias{damping_none}
\title{Do not dampen forecasts}
\usage{
damping_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option applies no damping to Rt forecasts, i.e. the trend
projected by the R estimation model will be continued until the end of the
forecast horizon.
}
