% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{dist_mean}
\alias{dist_mean}
\title{Get the mean of a discrete distribution}
\usage{
dist_mean(dist, include_zero = TRUE)
}
\arguments{
\item{dist}{Discrete distribution represented as numeric vector}

\item{include_zero}{If \code{TRUE} (default), the vector index starts at zero.
Otherwise, it starts at 1.}
}
\value{
Mean of the discrete distribution
}
\description{
Get the mean of a discrete distribution
}
