% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{dPCR_concentration_stepsize}
\alias{dPCR_concentration_stepsize}
\title{Calculate concentration step size from number of partitions and volume of a
digital PCR (dPCR) assay}
\usage{
dPCR_concentration_stepsize(partitions, volume_scaled)
}
\arguments{
\item{partitions}{The number of partitions in the dPCR assay.}

\item{volume_scaled}{The scaled volume of each partition (partition volume
multiplied with the scaling of concentration in the assay).}
}
\value{
The concentration step size.
}
\description{
This helper function calculates the concentration step
size implied by a certain number of partitions and the scaled volume of each
partition.
}
