% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{draws_to_init}
\alias{draws_to_init}
\title{Convert draws from a previous stan model fit to inits for MCMC sampling}
\usage{
draws_to_init(fit, model, num_chains = 4)
}
\arguments{
\item{fit}{A fitted model object (e.g. obtained from pathfinder).}

\item{model}{The model specification, used to identify scalar parameters.}

\item{num_chains}{The number of chains for which to generate inits.}
}
\value{
A list of inits for MCMC sampling, with one entry per chain.
}
\description{
Helper function that converts the draws obtained (e.g. from
pathfinder) to a list of inits that can be used for MCMC sampling.
}
\keyword{internal}
