% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{droplets_observe}
\alias{droplets_observe}
\title{Observe positive droplet counts from wastewater samples}
\usage{
droplets_observe(
  data,
  composite_window = 1,
  date_col = "date",
  droplets_col = "droplets",
  replicate_col = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicate_col}{}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Observe positive droplet counts from wastewater samples
}
\seealso{
Other {observation types}: 
\code{\link{concentrations_observe}()}
}
\concept{{observation types}}
