% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\docType{class}
\name{EpiSewerJob-class}
\alias{EpiSewerJob-class}
\title{The EpiSewerJob S4 class}
\description{
An S4 class registration for \code{EpiSewerJob} objects. Instances
are S3 objects created by \code{\link[=EpiSewerJob]{EpiSewerJob()}}. The class registration enables
\code{is(x, "EpiSewerJob")} checks.
}
\keyword{internal}
