% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\docType{class}
\name{EpiSewerJobResult-class}
\alias{EpiSewerJobResult-class}
\title{The EpiSewerJobResult S4 class}
\description{
An S4 class registration for \code{EpiSewerJobResult} objects.
Instances are S3 objects created by
\code{\link[=EpiSewerJobResult]{EpiSewerJobResult()}}. The class registration
enables \code{is(x, "EpiSewerJobResult")} checks.
}
\keyword{internal}
