% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{EpiSewerJobResult}
\alias{EpiSewerJobResult}
\title{Constructor for EpiSewerJobResult objects}
\usage{
EpiSewerJobResult(fit_res, job, stan_model, checksums)
}
\arguments{
\item{fit_res}{The result of model fitting, either a fitted model object or
an error object if model fitting failed.}

\item{job}{The EpiSewerJob object that was run.}

\item{stan_model}{The stan model that was used for fitting.}

\item{checksums}{A list with checksums that uniquely identify the model,
input, inits, fit_opts and results_opts of the job.}
}
\value{
An EpiSewerJobResult object with the results of model fitting and
relevant information about the job and the model.
}
\description{
Constructor for EpiSewerJobResult objects
}
