% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fence}
\alias{fence}
\title{Clip values of a vector between a lower and an upper bound}
\usage{
fence(vec, LB = -Inf, UB = Inf)
}
\arguments{
\item{vec}{A vector with values to be clipped}

\item{LB}{Lower bound}

\item{UB}{Upper bound}
}
\value{
A vector with clipped/fenced values
}
\description{
Clip values of a vector between a lower and an upper bound
}
