% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{find_mu_sigma_truncnorm}
\alias{find_mu_sigma_truncnorm}
\title{Find mu and sigma of a truncated normal distribution for given quantiles}
\usage{
find_mu_sigma_truncnorm(q5_target, q95_target)
}
\arguments{
\item{q5_target}{The target quantile at 5\% (lower bound).}

\item{q95_target}{The target quantile at 95\% (upper bound).}
}
\value{
A list containing the estimated \code{mu} and \code{sigma} for the truncated
normal distribution that matches the specified quantiles.
}
\description{
This function estimates the parameters \code{mu} and \code{sigma} of a
normal distribution truncated at zero, such that the 5\% and 95\% quantiles
of the distribution match the specified target quantiles.
}
\keyword{internal}
