% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{fit_model}
\alias{fit_model}
\title{Fit the EpiSewer model}
\usage{
fit_model(job, model, model_instance, run_silent = FALSE)
}
\arguments{
\item{job}{An EpiSewer job object containing the data, inits, and fitting
options.}

\item{model}{The model specification, used to identify parameters.}

\item{model_instance}{A compiled model object.}

\item{run_silent}{Should the fitting be run silently? Default is FALSE.}
}
\value{
A fitted model object if successful, or a list with error messages
and sampler output if there was an error during fitting.
}
\description{
Helper function that fits the EpiSewer model using the specified
sampler and model instance, with the data and inits from the job object.
}
\keyword{internal}
