% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit model via stan}
\usage{
fit_stan(stanmodel_instance, arguments, fit_method, silent = FALSE)
}
\arguments{
\item{stanmodel_instance}{A CmdStanModel object of the EpiSewer model
constructed using \code{cmdstanr::cmdstan_model()}.}

\item{arguments}{Arguments to be passed to stan, including the data and the
sampler options.}

\item{fit_method}{The algorithm used for sampling, currently supported are
"mcmc" and "pathfinder".}

\item{silent}{Should the sampling be completely silent or should status
updates and warnings be shown?}
}
\value{
Either a fitted stan model with draws, or a list with error messages
and sampler outputs.
}
\description{
Helper function to fit an EpiSewer model using one of stan's
algorithms.
}
\keyword{internal}
