% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{flatten_components}
\alias{flatten_components}
\title{Flatten modules/components into one slot-keyed component list}
\usage{
flatten_components(x)
}
\description{
Flatten modules/components into one slot-keyed component list
}
\keyword{internal}
