% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sewage.R
\name{flows_assume}
\alias{flows_assume}
\title{Assume a constant wastewater flow}
\usage{
flows_assume(flow_constant)
}
\arguments{
\item{flow_constant}{Daily wastewater flow volume, assumed to be the same for
all days.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option assumes a constant flow of wastewater at the
sampling site. Can be used as an approximation if no regular flow
measurements are available.
}
\details{
The flow volume unit should be the same as for the concentration
measurements, e.g. if concentrations are measured in gc/mL, then the flow
should be in mL as well.

Note that when the flow is unknown and therefore some arbitrary
value (e.g. \code{flow_constant=1}) is assumed, this must also be accounted for
in the assumed \code{load_per_case}. For this purpose, the function
\code{\link[=suggest_load_per_case]{suggest_load_per_case()}} offers a \code{flow_constant} argument.
}
\seealso{
Other flow models: 
\code{\link{flows_observe}()}
}
\concept{flow models}
