% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sewage.R
\name{flows_observe}
\alias{flows_observe}
\title{Observe wastewater flows}
\usage{
flows_observe(flows = NULL, date_col = "date", flow_col = "flow")
}
\arguments{
\item{flows}{A \code{data.frame} with each row representing one day. Must have at
least a column with dates and a column with flow measurements.}

\item{date_col}{Name of the column containing the dates.}

\item{flow_col}{Name of the column containing the flows.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option accounts for daily wastewater flow volumes measured
at the sampling site. The flow can change due to rainfall or industrial
discharge, and directly influences pathogen concentrations in the
wastewater.
}
\details{
The flow volume unit should be the same as for the concentration
measurements, e.g. if concentrations are measured in gc/mL, then the flow
should be in mL as well.
}
\seealso{
Other flow models: 
\code{\link{flows_assume}()}
}
\concept{flow models}
