% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{generation_dist_assume}
\alias{generation_dist_assume}
\title{Assume a generation time distribution}
\usage{
generation_dist_assume(generation_dist = NULL)
}
\arguments{
\item{generation_dist}{A numeric vector representing a discrete generation
time distribution, starting with the probability for a generation time of 1
day, 2 days, 3 days, and so on (a generation time of 0 days is excluded).}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option assumes a fixed generation time distribution for the
renewal model in \code{EpiSewer}.
}
\details{
The generation time distribution here refers to the intrinsic
distribution of the time between infection of a primary case and infection
of its secondary cases. It is disease-specific and typically obtained from
literature.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}},
\code{\link[=get_discrete_gamma_shifted]{get_discrete_gamma_shifted()}},
\code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
