% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_checksum_model}
\alias{get_checksum_model}
\title{Computes a checksum of a stanmodel, i.e. of the code from its main
stan file and all included stan files}
\usage{
get_checksum_model(stanmodel, only_functions = FALSE)
}
\description{
Computes a checksum of a stanmodel, i.e. of the code from its main
stan file and all included stan files
}
\keyword{internal}
