% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{get_checksums}
\alias{get_checksums}
\title{Get checksums that uniquely identify an EpiSewer job.}
\usage{
get_checksums(job, stanmodel = NULL)
}
\arguments{
\item{job}{An EpiSewer job object.}

\item{stanmodel}{Optional, a stanmodel. Can be passed to avoid recompilation
of the model. If NULL (default), the model defined in \code{job} is taken.}
}
\value{
A \code{list} with checksums identifying the model, input, inits and
fit_opts.
}
\description{
Get checksums that uniquely identify an EpiSewer job.
}
