% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_discrete_gamma}
\alias{get_discrete_gamma}
\title{Get PMF of a discretized Gamma distribution.}
\usage{
get_discrete_gamma(
  gamma_shape,
  gamma_rate,
  gamma_scale,
  gamma_mean,
  gamma_sd,
  maxX,
  include_zero = TRUE,
  print_params = FALSE
)
}
\arguments{
\item{gamma_shape}{Shape parameter of the Gamma distribution}

\item{gamma_rate}{Rate parameter of the Gamma distribution.}

\item{gamma_scale}{Scale parameter of the Gamma distribution. Can be
specified instead of the rate. Only has an effect if the rate is not
specified.}

\item{gamma_mean}{Alternative parameterization: Mean of the Gamma}

\item{gamma_sd}{Alternative parameterization: Standard deviation of the Gamma}

\item{maxX}{Right truncation point}

\item{include_zero}{Should the distribution explicitly cover X=0, or should
X=1 include the probability mass for X=0 too?}

\item{print_params}{Should the shape and rate parameters be printed?}
}
\value{
PMF of the discretized Gamma distribution
}
\description{
This function accepts different parameterizations to specify the Gamma
distribution
}
