% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_discrete_gamma_shifted}
\alias{get_discrete_gamma_shifted}
\title{Title}
\usage{
get_discrete_gamma_shifted(gamma_mean = 4.8, gamma_sd = 2.3, maxX = 10)
}
\arguments{
\item{maxX}{}
}
\value{

}
\description{
Title
}
\section{This ocode is adapted from EpiEstim, credit to Anne Cori}{
a.cori@imperial.ac.uk
}

