% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_discrete_lognormal}
\alias{get_discrete_lognormal}
\title{Get PMF of a discretized lognormal distribution}
\usage{
get_discrete_lognormal(
  meanlog,
  sdlog,
  unit_mean = NULL,
  unit_sd = NULL,
  unit_cv = NULL,
  maxX = NULL,
  include_zero = TRUE,
  print_params = FALSE
)
}
\arguments{
\item{meanlog}{Log scale mean (location of lognormal).}

\item{sdlog}{Log scale standard deviation (scale of lognormal).}

\item{unit_mean}{Alternative parameterization: unit/natural scale mean.}

\item{unit_sd}{Alternative parameterization: unit/natural scale sd.}

\item{unit_cv}{Alternative parameterization: unit/natural scale coefficient
of variation.}

\item{maxX}{Right truncation point. All probability mass beyond \code{maxX} will
be assigned to \code{maxX}. If \code{NULL} (default), this is automatically chosen
such that the last bin has less than 0.5\% of the probability mass.}

\item{include_zero}{Should the distribution explicitly cover X=0, or should
X=1 include the probability mass for X=0 too?}

\item{print_params}{Should the log-level parameters be printed?}
}
\value{
A numeric vector representing the PMF of the discretized lognormal.
}
\description{
This function accepts both log-scale and unit-scale parameters to specify a
discretized lognormal distribution.
}
