% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_from_env}
\alias{get_from_env}
\title{Try to get an object from the upper environments on the calling stack}
\usage{
get_from_env(obj, arg = NULL, max_levels = 10)
}
\description{
Try to get an object from the upper environments on the calling stack
}
