% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_gamma_mean_alternative}
\alias{get_gamma_mean_alternative}
\title{Get mean of a Gamma distribution given its shape and scale}
\usage{
get_gamma_mean_alternative(gamma_shape, gamma_scale)
}
\description{
Get mean of a Gamma distribution given its shape and scale
}
