% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_gamma_rate_alternative}
\alias{get_gamma_rate_alternative}
\title{Get rate of a Gamma distribution given its mean and sd}
\usage{
get_gamma_rate_alternative(gamma_mean, gamma_sd)
}
\description{
Get rate of a Gamma distribution given its mean and sd
}
