% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_gamma_scale_alternative}
\alias{get_gamma_scale_alternative}
\title{Get scale of a Gamma distribution given its mean and sd}
\usage{
get_gamma_scale_alternative(gamma_mean, gamma_sd)
}
\description{
Get scale of a Gamma distribution given its mean and sd
}
