% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_gamma_sd_alternative}
\alias{get_gamma_sd_alternative}
\title{Get sd of a Gamma distribution given its shape and scale}
\usage{
get_gamma_sd_alternative(gamma_shape, gamma_scale)
}
\description{
Get sd of a Gamma distribution given its shape and scale
}
