% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_hash_model}
\alias{get_hash_model}
\title{Computes a deterministic hash of a stanmodel, i.e. of the code from its main
stan file and all included stan files}
\usage{
get_hash_model(stanmodel, only_functions = FALSE)
}
\description{
Computes a deterministic hash of a stanmodel, i.e. of the code from its main
stan file and all included stan files
}
