% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_calibrate.R
\name{get_infection_curve_crude}
\alias{get_infection_curve_crude}
\title{Compute crude estimate of infections by date of infections}
\usage{
get_infection_curve_crude(load_curve_crude, load_per_case)
}
\arguments{
\item{load_curve_crude}{A \code{data.frame} with columns \code{date_index}, \code{date}, and
\code{load} as produced by \code{\link[=get_load_curve_crude]{get_load_curve_crude()}}.}

\item{load_per_case}{Load per case in gene copies.}
}
\value{
A \code{data.frame} with columns \code{date_index}, \code{date}, and \code{load}.
}
\description{
The function smoothed the measured loads using a loess smoother
on the log scale and shifts them backward by the mean total delay to obtain
a crude estimate of the loads by date of infection. This is useful for
different calibration purposes, such as determining a suitable
\code{load_per_case} assumption.
}
