% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_lognormal_mu_alternative}
\alias{get_lognormal_mu_alternative}
\title{Compute mu parameter of Log-Normal from other quantities.}
\usage{
get_lognormal_mu_alternative(
  unit_mean = NULL,
  unit_sd = NULL,
  unit_q5 = NULL,
  unit_q95 = NULL
)
}
\arguments{
\item{unit_q5}{5\% quantile of distribution.}

\item{unit_q95}{95\% quantile of distribution.}
}
\value{
Mu parameter of Log-Normal distribution.
}
\description{
Compute mu parameter of Log-Normal from other quantities.
}
\details{
Currently, only conversion from quantiles is supported, but other
alternatives may be added.
}
