% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_lognormal_sigma_alternative}
\alias{get_lognormal_sigma_alternative}
\title{Compute sigma parameter of Log-Normal from other quantities.}
\usage{
get_lognormal_sigma_alternative(
  mu,
  unit_mean = NULL,
  unit_sd = NULL,
  unit_q5 = NULL,
  unit_q95 = NULL
)
}
\arguments{
\item{mu}{Mu parameter of Log-Normal distribution.}

\item{unit_q5}{5\% quantile of distribution.}

\item{unit_q95}{95\% quantile of distribution.}
}
\value{
Sigma parameter of Log-Normal distribution.
}
\description{
Compute sigma parameter of Log-Normal from other quantities.
}
\details{
Currently, only conversion from mu + a quantile is supported, but
other alternatives may be added.
}
