% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_pathfinder_inits}
\alias{get_pathfinder_inits}
\title{Run stan's pathfinder algorithm to obtain improved inits.}
\usage{
get_pathfinder_inits(stanmodel_instance, job)
}
\arguments{
\item{stanmodel_instance}{A CmdStanModel object of the EpiSewer model
constructed using \code{cmdstanr::cmdstan_model()}.}

\item{job}{An EpiSewer job object with data and default inits.}
}
\value{
Fitted model, a CmdStanPathfinder object. Can be used as arguments
for the inits of the \code{CmdStanModel$sample()} method.
}
\description{
Helper function that runs the pathfinder algorithm with default
inits and returns the result to be used as inits in mcmc sampling.
}
\keyword{internal}
