% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_regression_linear_trend}
\alias{get_regression_linear_trend}
\title{Obtain regression vector to estimate a linear trend}
\usage{
get_regression_linear_trend(x, weights = rep(1/length(x), length(x)))
}
\arguments{
\item{x}{A \code{vector} with observations of the single independent variable (for
example points in time).}

\item{weights}{Optional, a \code{vector} with weights for each observation.}
}
\value{
A \code{vector} which contains the relevant row of the regression solution
matrix that represents the trend coefficient. By multiplying this vector
with the observed time series values y, you directly get the (weighted
least squares) estimate of the trend.

#@examples
#x = 1:6
#y = x\emph{4 + rnorm(6, 0, 0.1)
#w <- c(0.1, 0.1, 0.1, 0.2, 0.2, 0.3) # weights
#trend_reg <- get_regression_linear_trend(x, weights = w)
#as.vector(trend_reg \%}\% y) # this gives you the trend estimate
#summary(lm(y ~ x, weights = w))$coefficients\link{"x","Estimate"} # should be the same
}
\description{
This function provides the solution vector from a (weighted)
linear regression with one independent variable x, where x may represent
points in a time series.
}
\keyword{internal}
