% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_stan_files_digest}
\alias{get_stan_files_digest}
\title{Computes a digest of the stan model source files, i.e. of the main stan
file and all included stan files (without requiring the model to be
compiled)}
\usage{
get_stan_files_digest(
  modelfolder = "inst/stan",
  modelname = "EpiSewer_main.stan"
)
}
\description{
Computes a digest of the stan model source files, i.e. of the main stan
file and all included stan files (without requiring the model to be
compiled)
}
\keyword{internal}
