% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Load stan model}
\usage{
get_stan_model(
  modeldata = NULL,
  model_filename = NULL,
  model_folder = "stan",
  profile = TRUE,
  threads = FALSE,
  force_recompile = FALSE,
  package = "EpiSewer"
)
}
\arguments{
\item{modeldata}{A \code{modeldata} object containing data and specifications of
the model to be fitted. The required stan model is automatically inferred
from the \code{modeldata}.}

\item{model_filename}{File name of a specific stan model to load. This is an
alternative to supplying \code{modeldata}.}

\item{model_folder}{Path to the folder containing the stan models for
\code{EpiSewer}.}

\item{profile}{Should profiling be run during model fitting? Default is
\code{TRUE}. If \code{FALSE}, will remove all profiling statements from the model
before fitting. This can decrease runtime in some cases.}

\item{threads}{Should multihreading be enabled? Default is \code{FALSE}, as
\code{EpiSewer} currently does not support within-chain parallelism.}

\item{force_recompile}{If \code{FALSE} (default), the model is only recompiled if
changes to the model code are detected. However, as the change detection is
not fully reliable, it is sometimes necessary to force recompilation after
having made some changes to the stan code.}

\item{package}{Name of the package in which to search for stan files
(defaults to EpiSewer). If NULL, will search in the normal working directory.}
}
\value{
A \code{list} containing the model definition and a link to the compiled
stan model.
}
\description{
Loads a specific stan model for model fitting in \code{EpiSewer}.
}
