% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{horizon_assume}
\alias{horizon_assume}
\title{Specify the forecast horizon}
\usage{
horizon_assume(horizon)
}
\arguments{
\item{horizon}{The forecast horizon in days. If 0, no forecasts are produced.
Note that this functionality is intended for short-term forecasts.
Projections over longer horizons can be highly inaccurate.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option specifies a fixed forecast horizon in days. EpiSewer
will produce concentration predictions and forecasts of all latent
variables such as Rt, infections and load until the end of the forecast
horizon.
}
\details{
Forecasts account for the estimated variation of transmission
dynamics over time and therefore tend to become more uncertain at longer
forecast horizons. However, it is important to keep in mind that depending
on the Rt model used, EpiSewer will project the current transmission
dynamics to continue unchanged (when using \code{\link[=R_estimate_rw]{R_estimate_rw()}},
\code{\link[=R_estimate_splines]{R_estimate_splines()}}, \code{\link[=R_estimate_piecewise]{R_estimate_piecewise()}}) or according to a
(dampened) linear trend (when using \code{\link[=R_estimate_ets]{R_estimate_ets()}} and
\code{\link[=R_estimate_changepoint_splines]{R_estimate_changepoint_splines()}}). This assumption can be violated by
various factors such as depletion of susceptible individuals, changes in
behavior, or public health interventions.
}
