% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{horizon_none}
\alias{horizon_none}
\title{Do not produce forecasts}
\usage{
horizon_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option specifies that no forecasts should be made. Only
estimates and concentration predictions until the last observed date are
produced.
}
