% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{incubation_dist_assume}
\alias{incubation_dist_assume}
\title{Assume an incubation period distribution}
\usage{
incubation_dist_assume(incubation_dist = NULL)
}
\arguments{
\item{incubation_dist}{A numeric vector representing a discrete incubation
period distribution, starting with the probability for an incubation period
of 0 days, 1 day, 2 days, and so on.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option assumes a fixed incubation period distribution for
the shedding model in \code{EpiSewer}.
}
\details{
The incubation period is the time between infection and symptom
onset. This assumption is used when \code{shedding_reference="symptom_onset"},
i.e. to support shedding load distributions referenced by days since
symptom onset.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
