% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{infection_noise_none}
\alias{infection_noise_none}
\title{Do not model infection noise}
\usage{
infection_noise_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option does not model noise in the infection process and instead
implements a deterministic renewal model.
}
\seealso{
Other infection noise models: 
\code{\link{infection_noise_estimate}()}
}
\concept{infection noise models}
