% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Influenza_A_Zurich}
\alias{Influenza_A_Zurich}
\title{Raw Influenza A wastewater data from Zurich}
\format{
A list with the following elements:
\describe{
  \item{measurements}{A \code{data.table} with columns \code{date},
    \code{concentration} (gc/mL), \code{replicate_id}, \code{dilution},
    \code{total_droplets}, and \code{positive_droplets}.}
  \item{flows}{A \code{data.table} with columns \code{date} and \code{flow} (mL/day).}
  \item{units}{A list with concentration and flow unit strings.}
}
}
\description{
Wastewater measurements of Influenza A RNA from the WWTP
Werdhölzli in Zürich, Switzerland (September 2022–April 2023). Data are
provided by EAWAG.
}
\keyword{datasets}
\keyword{internal}
