% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{init_from_normal_prior}
\alias{init_from_normal_prior}
\title{Provide initialization value for a parameter based on the supplied prior}
\usage{
init_from_normal_prior(prior)
}
\arguments{
\item{prior}{Prior for parameter as provided by \code{\link[=set_prior]{set_prior()}}. Should be a
normal prior.}
}
\value{
Mean of the normal prior plus 1/4 of the standard deviation.
}
\description{
Initialization using the prior is often better than initializing
with zero (and if the parameter is strictly positive, zero is not possible
at all). This function provides as init value the mean of the normal prior
plus 1/4 of the standard deviation. This ensure a positive init even if the
mean is zero (useful for truncated normal priors.)
}
