% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{load_variation_none}
\alias{load_variation_none}
\title{Do not model individual-level load variation}
\usage{
load_variation_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option assumes that there is no variation in the
total load shed per case, i.e. the individual shedding load is fixed to the
average shedding load.
}
\seealso{
Other load variation models: 
\code{\link{load_variation_estimate}()}
}
\concept{load variation models}
