% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_assume}
\alias{LOD_assume}
\title{Assume a limit of detection}
\usage{
LOD_assume(
  limit = NULL,
  prob = 0.95,
  LOD_type = "exponential",
  drop_prob = 1e-10
)
}
\arguments{
\item{limit}{Limit of detection. The concentration below which the pathogen
cannot be detected with sufficient probability, i.e. the measurement may be
zero although the pathogen is present in the sample.}

\item{prob}{What desired probability of detection does the limit refer to?
Default is 95\% (0.95): This means that the provided \code{limit} is the smallest
concentration at which the pathogen can still be detected with over 95\%
probability.}

\item{LOD_type}{The type of LOD model used. Currently, only "exponential" is
supported. This models an exponentially decreasing probability of zero
measurements / non-detection as a function of concentration. The
exponential model can be derived from the statistical properties of dPCR,
but should also work well for other quantification methods such as qPCR.}

\item{drop_prob}{Probability for non-detection below which likelihood
contributions of observed concentrations are dropped from LOD model. This
avoids numerical issues of the LOD model at high concentrations (very small
non-detection probabilities) that can otherwise affect sampling speed.
Since these likelihood contributions will be virtually zero for almost all
samples anyway, parameter estimates are practically not affected.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Pathogen concentrations below a certain threshold may not be
detectable and thus erroneously measured as 0. This option adjusts for a
known limit of detection and includes zero measurements in the likelihood.
}
\details{
The limit of detection is modeled using a hurdle model. In effect,
zero measurements provide a signal that the concentration in the respective
sample was likely below the limit of detection, but we don't know what the
exact concentration was.

The limit of detection is specific to the quantification approach
and protocol. It is usually established from a dedicated lab experiment
(serial dilution experiment). It his here assumed that this experiment did
not cover a large fraction of the preprocessing noise to find an optimal
configuration for the exponential model.

If used together with \code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}} or
\code{\link[=noise_estimate_dPCR_params]{noise_estimate_dPCR_params()}}, EpiSewer will also model the effect of
pre-PCR noise on the LOD. This means that the modeled LOD could be slightly
higher than specified under \code{limit}, depending on the estimated pre-PCR
noise.
}
\seealso{
{Visualize the assumed LOD as a function of concentration:}
\code{\link[=plot_LOD]{plot_LOD()}}

Other LOD models: 
\code{\link{LOD_estimate_dPCR}()},
\code{\link{LOD_none}()}
}
\concept{LOD models}
