% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_estimate_dPCR}
\alias{LOD_estimate_dPCR}
\title{Estimate a limit of detection model for digital PCR data}
\usage{
LOD_estimate_dPCR(drop_prob = 1e-10)
}
\arguments{
\item{drop_prob}{Probability for non-detection below which likelihood
contributions of observed concentrations are dropped from LOD model. This
avoids numerical issues of the LOD model at high concentrations (very small
non-detection probabilities) that can otherwise affect sampling speed.
Since these likelihood contributions will be virtually zero for almost all
samples anyway, parameter estimates are practically not affected.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Pathogen concentrations below a certain threshold may not be
detectable and thus erroneously measured as 0. This option adjusts for a
limit of detection based on the statistical properties of digital PCR
(dPCR) and includes zero measurements in the likelihood.

In effect, zero measurements provide a signal that the
concentration in the respective sample was likely below the limit of
detection, but we don't know what the exact concentration was.
}
\details{
The limit of detection is modeled using a hurdle model. The model
uses the number of partitions in the dPCR reaction and the conversion
factor as defined and estimated by \code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}}. It can
therefore only be used together with \code{noise = noise_estimate_dPCR()} or
\code{noise = noise_estimate_dPCR_params()} in \code{\link[=model_measurements]{model_measurements()}}.
}
\seealso{
Other LOD models: 
\code{\link{LOD_assume}()},
\code{\link{LOD_none}()}
}
\concept{LOD models}
