% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_none}
\alias{LOD_none}
\title{Do not model a limit of detection}
\usage{
LOD_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option drops all zero measurements from the likelihood and
does not explicitly model a limit of detection.
}
\details{
Dropping zero measurements implicitly assumes that any
level of concentration can theoretically lead to a zero measurement. Since
the corresponding observations are dropped, this will discard information,
but not bias estimates.
}
\seealso{
Other LOD models: 
\code{\link{LOD_assume}()},
\code{\link{LOD_estimate_dPCR}()}
}
\concept{LOD models}
