% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{match_obs_dist}
\alias{match_obs_dist}
\title{Check the supplied observation distribution type and return the internal id}
\usage{
match_obs_dist(distribution)
}
\description{
Check the supplied observation distribution type and return the internal id
}
\keyword{internal}
