% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mdaccess.R
\name{md_can}
\alias{md_can}
\title{Check whether a modeldata variable is present or derivable}
\usage{
md_can(md, path, visiting = character(0))
}
\description{
Check whether a modeldata variable is present or derivable
}
\keyword{internal}
