% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mdaccess.R
\name{md_derivation_index}
\alias{md_derivation_index}
\title{Get (and lazily build) the provides-path -> derivation-unit index}
\usage{
md_derivation_index(md)
}
\description{
Get (and lazily build) the provides-path -> derivation-unit index
}
\keyword{internal}
