% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mdaccess.R
\name{md_has_path}
\alias{md_has_path}
\title{Check presence of a modeldata variable by \code{$}-path}
\usage{
md_has_path(md, path)
}
\description{
Check presence of a modeldata variable by \code{$}-path
}
\keyword{internal}
