% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mdaccess.R
\name{md_set_path}
\alias{md_set_path}
\title{Set a (possibly nested) modeldata variable by \code{$}-path}
\usage{
md_set_path(md, path, value)
}
\description{
Set a (possibly nested) modeldata variable by \code{$}-path
}
\keyword{internal}
