% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{measurement_noise_estimate}
\alias{measurement_noise_estimate}
\title{Title}
\usage{
measurement_noise_estimate(
  sigma_prior = set_prior("sigma", "normal", mu = 0, sigma = 1),
  modeldata = modeldata_init()
)
}
\arguments{
\item{sigma_prior}{}
}
\value{

}
\description{
Title
}
