% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelcomponent.R
\docType{class}
\name{model_component-class}
\alias{model_component-class}
\title{The model_component S4 class registration}
\description{
The model_component S4 class registration
}
\keyword{internal}
