% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{model_stan_opts}
\alias{model_stan_opts}
\title{Specify details of the stan model}
\usage{
model_stan_opts(
  model_filename = NULL,
  model_folder = "stan",
  profile = TRUE,
  threads = FALSE,
  force_recompile = FALSE,
  package = "EpiSewer"
)
}
\arguments{
\item{model_filename}{Name of the stan model file. If NULL (default), the
name of the model file is automatically inferred from \code{modeldata}.}

\item{model_folder}{(Relative) path to the folder containing the stan model.}

\item{profile}{Should profiling be run during model fitting? Default is
\code{TRUE}. If \code{FALSE}, will remove all profiling statements from the model
before fitting. This can decrease runtime in some cases.}

\item{threads}{Should multihreading be enabled? Default is \code{FALSE}, as
\code{EpiSewer} currently does not support within-chain parallelism.}

\item{force_recompile}{Should recompilation be forced before model fitting?
If \code{FALSE} (default), the model is only recompiled when changes to the
model code are detected. However, as the change detection is not fully
reliable, it is sometimes necessary to force recompilation after having
made changes to the stan code.}

\item{package}{Name of the package in which to search for stan files
(defaults to EpiSewer). If NULL, will search in the normal working
directory, allowing users to provide their own customized model files.}
}
\value{

}
\description{
Specify details of the stan model
}
