% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{modeldata_check}
\alias{modeldata_check}
\title{Check modeldata object for required variables}
\usage{
modeldata_check(
  modeldata,
  required,
  descriptions = modeldata_descriptions(),
  run_before = modeldata_var_requirements(),
  advice = NULL,
  throw_error = TRUE,
  calling_env = rlang::caller_env()
)
}
\arguments{
\item{modeldata}{A \code{list} with all data variables (including priors) to be
passed on to stan, alongside additional meta information and descriptions.}

\item{required}{Character vector of required variables}

\item{descriptions}{Optional, a list with names corresponding to all or some
of the required variables, and values corresponding to additional
descriptions of the variables}

\item{run_before}{Optional, either a character vector, or a \code{list} with names
corresponding to all of the required variables, and values corresponding to
functions that will add these variables to modeldata.}

\item{advice}{Additional message with advice on how to solve error.}

\item{throw_error}{Should an error be thrown (with verbatim error message),
or should \code{TRUE}/\code{FALSE} be returned. Default is \code{TRUE}.}
}
\value{
If throw_error is \code{TRUE}, nothing is returned, otherwise a logical is
returned (\code{TRUE} if all required variables are present)
}
\description{
\code{modeldata_check} accepts various additional arguments which are turned into
a sophisticated error message about missing variables
}
