% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{modeldata_init}
\alias{modeldata_init}
\title{Construct an unspecified EpiSewer model}
\usage{
modeldata_init()
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
Construct an unspecified EpiSewer model
}
