% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{modeldata_validate}
\alias{modeldata_validate}
\title{Ensure that modeldata contains all necessary variables}
\usage{
modeldata_validate(modeldata, model_def, defaults = modeldata_defaults())
}
\arguments{
\item{modeldata}{A \code{list} with all data variables (including priors) to be
passed on to stan, alongside additional meta information and descriptions.}

\item{model_def}{A \code{list} with information about the stan model to be fitted
and a function that returns the CmdStanModel object.}

\item{defaults}{A \code{list} with default values to be used for modeldata
variables if not supplied in modeldata. For example, \code{numeric(0)} will often
be supplied for optional parameters.}
}
\value{
If no error is thrown due to missing mandatory variables, the same
modeldata object is returned again, where optional variables have been
replaced by zero-length dimension defaults for stan.
}
\description{
Throws an error if necessary variables are missing, and replaces optional
variables with defaults indicating their absence in the stan model.
}
\details{
The reason why defaults are inserted for missing optional variables is that
stan does not explicitly allow default variables, so they are specified with
a little hack by using zero-length dimensions.
}
