% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_}
\alias{noise_estimate_}
\title{Estimate measurement noise (internal helper function)}
\usage{
noise_estimate_(
  replicates = FALSE,
  distribution,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  cv_type = "constant",
  total_partitions_observe = NULL,
  max_partitions_prior_lower = NULL,
  max_partitions_prior_upper = NULL,
  partition_loss_mean_prior_lower = NULL,
  partition_loss_mean_prior_upper = NULL,
  partition_loss_variation_prior_lower = NULL,
  partition_loss_variation_prior_upper = NULL,
  partition_loss_max = NULL,
  volume_scaled_prior_lower = NULL,
  volume_scaled_prior_upper = NULL,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  prePCR_noise_type = "log-normal",
  use_taylor_approx = TRUE,
  helper_name = "noise_estimate"
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that when \code{replicates=TRUE}, this is only
the CV after the replication stage (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{cv_type}{One out of "constant" (default), "constant_var", or "dPCR". If
"constant", the coefficient of variation is estimated as a constant/single
parameter for all observations. If "dPCR", the coefficient of variation is
modeled as a function of the expected concentration according to the
statistical properties of dPCR. In particular, this model predicts a higher
coefficient of variation at smaller concentrations, which often leads to a
better model fit. If "constant_var", not the coefficient of variation but
the variance of measurements is modeled as constant. This is usually a
misspecification and is only supported for comparison purposes.}

\item{total_partitions_observe}{If TRUE, the total number of partitions is
taken from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_partitions_col} is specified in \code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{max_partitions_prior_lower}{Prior (lower bound) for the maximum total
number of dPCR partitions. This is usually defined by the manufacturer of
the dPCR system/chip used, which supports a certain maximum number of
partitions. If you know the exact dPCR system and its maximum partition
number, you can set both \code{max_partitions_prior_lower} and
\code{max_partitions_prior_upper} to this value. Otherwise, this prior can be
used to set a broad lower and upper bound for the maximum number of
partitions, to reflect a range of popular dPCR systems/chips.}

\item{max_partitions_prior_upper}{Prior (upper bound) for the maximum total
number of dPCR partitions (see \code{max_partitions_prior_lower} for details.)}

\item{partition_loss_mean_prior_lower}{Prior (5\% quantile) for the mean
relative partition loss. A certain proportion of partitions in a dPCR run
is typically invalid and discarded from the concentration estimate. This
prior can be used to set a lower and upper bound for the mean proportion of
partitions lost. Note that for proportions close to 0 or to
\code{partition_loss_max} (see below), the resulting mean partition loss can
differ from what is specified here, because we internally
translate this prior to the logit scale.}

\item{partition_loss_mean_prior_upper}{Prior (95\% quantile) for the mean
relative partition loss (see \code{partition_loss_mean_prior_lower} for
details). In well-functioning dPCR assays, the mean proportion of lost
partitions should not be very high (definitely below 50\%).}

\item{partition_loss_variation_prior_lower}{Prior (5\% quantile) for the
variation in the number of invalid partitions across dPCR runs. The
proportion of partitions lost typically varies between dPCR runs. We thus
model this proportion as logit-normal distributed, with mean
(approximately) defined by \code{partition_loss_mean_prior} and logit-level
standard deviation \code{sigma}. You can use
\code{partition_loss_variation_prior_lower} and
\code{partition_loss_variation_prior_upper} to set a lower and upper bound for
\code{sigma}.}

\item{partition_loss_variation_prior_upper}{Prior (95\% quantile) for the
variation in the number of invalid partitions across dPCR runs (see
\code{partition_loss_variation_prior_lower} for details).}

\item{partition_loss_max}{The maximum proportion of partitions that can be
lost in a valid dPCR run. During quality control, runs where the proportion
of invalid partitions is above some threshold (e.g. 50\%) are often
discarded. This parameter can be used to represent such a QC threshold.}

\item{volume_scaled_prior_lower}{Prior (lower bound) on the conversion factor
(partition volume scaled by the dilution of wastewater in the assay) for
the dPCR reaction. See details for further explanation.}

\item{volume_scaled_prior_upper}{Prior (upper bound) on the conversion factor
(partition volume scaled by the dilution of wastewater in the assay) for
the dPCR reaction. If this is identical to \code{volume_scaled_prior_lower}, the
conversion factor will be fixed and not estimated.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} is used.}

\item{use_taylor_approx}{If TRUE (default), a Taylor expansion approximation
is used to estimate the CV of measurements under pre-PCR noise. The
approximation is very accurate, unless concentrations are extremely high
(so high that the quality of the measurements from dPCR would anyway be
questionable).}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option estimates the unexplained variation in wastewater
measurements. If multiple measurements (replicates) per sample are
provided, \code{EpiSewer} can also explicitly model variation before the
replication stage.

This helper function is called from other noise modeling
functions. \code{\link[=noise_estimate]{noise_estimate()}} is a constant coefficient of variation model,
\code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}} and \code{\link[=noise_estimate_dPCR_params]{noise_estimate_dPCR_params()}} are noise models
specialized for digital PCR (\code{cv_type = "dPCR"}), which may however also
work with other quantification methods such as qPCR, and
\code{\link[=noise_estimate_constant_var]{noise_estimate_constant_var()}} is a constant variance model.
}
\keyword{internal}
