% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_constant_var}
\alias{noise_estimate_constant_var}
\title{Estimate measurement noise with constant variance}
\usage{
noise_estimate_constant_var(
  replicates = FALSE,
  distribution = "gamma",
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  warn = TRUE
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{distribution}{Parametric distribution for measurement noise.
Currently supported are "gamma" (default and recommended), "log-normal",
"truncated normal", and "normal". The "truncated normal" and "normal"
options are not recommended for use in practice.}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that when \code{replicates=TRUE}, this is only
the CV after the replication stage (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{warn}{If \code{TRUE} (default), a warning is shown that constant variance
modeling is not recommended. Set to \code{FALSE} to suppress this warning.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a constant variance model. This is usually a
misspecification and is only supported for comparison purposes.

For a constant coefficient of variation model, see
\code{\link[=noise_estimate]{noise_estimate()}}, and for a non-constant coefficient of variation model,
see \code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}}.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
Note that although this model keeps the variance constant, the prior
for the measurement noise is still in terms of the (average) coefficient of
variation (CV). This makes prior specification easier since the CV is
unitless.

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage:
\verb{Truncated normal}
}
}
\seealso{
Other noise models: 
\code{\link{noise_estimate}()},
\code{\link{noise_estimate_dPCR}()},
\code{\link{noise_estimate_dPCR_params}()}
}
\concept{noise models}
