% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_ddPCR}
\alias{noise_estimate_ddPCR}
\title{Estimate measurement noise for digital droplet PCR data}
\usage{
noise_estimate_ddPCR(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  ddPCR_prior_droplets_mu = 20000,
  ddPCR_prior_droplets_sigma = 5000,
  ddPCR_droplets_fixed = TRUE,
  ddPCR_droplets_observe = FALSE,
  ddPCR_prior_scaling_mu = 1.5e-05,
  ddPCR_prior_scaling_sigma = 5e-06,
  ddPCR_scaling_fixed = FALSE,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  prePCR_noise_type = "log-normal",
  use_taylor_approx = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that in contrast to using
\code{\link[=noise_estimate]{noise_estimate()}}, this does \emph{not} include the technical noise of the
digital PCR. This is because the dPCR noise is explicitly modeled (using
the number of partitions and conversion factor). Moreover, when
\code{replicates=TRUE}, this is only the CV after the replication stage (see
details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{ddPCR_prior_droplets_mu}{Prior (mean) on the total number of droplets
in the ddPCR reaction.}

\item{ddPCR_prior_droplets_sigma}{Prior (standard deviation) on the total
number of droplets in the ddPCR reaction.}

\item{ddPCR_droplets_fixed}{If TRUE (default), the number of total droplets
is fixed to the prior mean and not estimated. This is recommended if no
replicates are available.}

\item{ddPCR_droplets_observe}{If TRUE, the number of total droplets is taken
from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_droplets_col} is specified in
\code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{ddPCR_prior_scaling_mu}{Prior (mean) on the concentration scaling
factor for the ddPCR reaction. The concentration scaling factor is the
droplet volume, scaled by the dilution of the wastewater in the ddPCR
reaction. See details for further explanation.}

\item{ddPCR_prior_scaling_sigma}{Prior (standard deviation) on the
concentration scaling factor for the ddPCR reaction.}

\item{ddPCR_scaling_fixed}{If TRUE, the concentration scaling factor is fixed
to the prior mean and not estimated.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=LOD_estimate_ddPCR]{LOD_estimate_ddPCR()}} is used.}

\item{use_taylor_approx}{If TRUE (default), a Taylor expansion approximation
is used to estimate the CV of measurements under pre-PCR noise. The
approximation is very accurate, unless concentrations are extremely high
(so high that the quality of the measurements from ddPCR would anyway be
questionable).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a coefficient of variation model specialised for digital
droplet PCR. Specifically, the coefficient of variation is modeled as a
function of the expected concentration according to the statistical
properties of ddPCR.

This model predicts a higher coefficient of variation at smaller
concentrations, which often leads to a better model fit, even for
measurements from other quantification methods such as qPCR.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
The concentration scaling factor (see \code{ddPCR_prior_scaling_mu},
\code{ddPCR_prior_scaling_sigma}, \code{ddPCR_scaling_fixed}) is the droplet volume,
scaled by the dilution of the wastewater in the ddPCR reaction. The
dilution accounts for all extraction and preparation steps. For example, if
the droplet volume is 4.5e-7 mL and the dilution of the wastewater is 100:3
(i.e. 100 gc/mL in the original wastewater sample correspond to 3 gc/mL in
the PCR reaction), then the overall scaling factor is 4.5e-7 * 100 / 3 =
1.5e-5.

When \code{replicates=TRUE}, two coefficients of variation are estimated:
\itemize{
\item the CV before the replication stage (see \code{pre_replicate_cv_prior_mu})
\item the CV after the replication stage (see \code{cv_prior_mu})
}

The meaning of these CV estimates depends on the type of replicates. If the
replicates are biological replicates (i.e. independently processed), then
\code{cv} estimates the noise in the preprocessing before the PCR, and
\code{pre_replicate_cv} estimates the noise from anything before preprocessing
(e.g. sampling noise and all other unexplained variation). In contrast, if
the replicates are technical replicates (i.e. several PCR runs of the same
preprocessed sample), then \code{cv} estimates only unexplained PCR noise
(should be close to zero), and \code{pre_replicate_cv} estimates all other noise
(including preprocessing noise.)

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
\item prior for number of droplets in ddPCR: \verb{Truncated normal}
\item prior for concentration scaling factor of ddPCR: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage (\code{pre_replicate_cv}):
\verb{Truncated normal}
}
}
\seealso{
\link{LOD_estimate_ddPCR} for a limit of detection model specialised for
ddPCR.

Other \link{noise models}: 
\code{\link{noise_estimate_constant_var}()},
\code{\link{noise_estimate}()}
}
\concept{\link{noise models}}
