% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_dPCR}
\alias{noise_estimate_dPCR}
\title{Estimate measurement noise for digital PCR data}
\usage{
noise_estimate_dPCR(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  total_partitions_observe = FALSE,
  max_partitions = 30000,
  partition_loss_mean = 0.1,
  volume_scaled = 1e-05,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  prePCR_noise_type = "log-normal"
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of analyzed
concentrations (pre-PCR noise). Note that this does \emph{not} include the
technical noise of the digital PCR. This is because the dPCR noise is
explicitly modeled (using the total number of partitions and conversion
factor). Moreover, when \code{replicates=TRUE}, noise before the replication
stage is separately modeled, so \code{cv_prior_mu} only describes any
remaining pre-PCR noise (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of analyzed concentrations.}

\item{total_partitions_observe}{If TRUE, the total number of partitions is
taken from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_partitions_col} is specified in \code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{max_partitions}{The maximum total number of dPCR partitions. This is
usually defined by the manufacturer of the dPCR system/chip used, which
supports a certain maximum number of partitions.}

\item{partition_loss_mean}{The mean relative partition loss. A certain
proportion of partitions in a dPCR run is typically invalid and discarded
from the concentration estimate.}

\item{volume_scaled}{The conversion factor (partition volume multiplied with
the scaling of concentration in the assay) for the dPCR reaction. See
details for further explanation.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} is used.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option models concentration measurements using an error
model specialized for digital PCR (e.g. digital droplet PCR). This is a
fast approximation assuming a fixed number of valid total partitions in the
dPCR assay and a known conversion factor. For a model that jointly
estimates uncertain assay parameters, see \code{\link[=noise_estimate_dPCR_params]{noise_estimate_dPCR_params()}}.

This model predicts more relative variation at smaller
concentrations, which often leads to a better model fit, even for
measurements from other quantification methods such as qPCR.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
The conversion factor (see \code{volume_scaled}) is the partition volume
v multiplied with a scaling factor s. The scaling factor accounts for
concentration differences between the sample and the reaction mix, for
example due to extraction or adding of reagents. For example, if the
partition volume is 4.5e-7 mL and the scaling factor is 100:3 (i.e. 100
gc/mL in the original sample correspond to 3 gc/mL in the PCR reaction),
then the overall conversion factor is 4.5e-7 * 100 / 3 = 1.5e-5.

When \code{replicates=TRUE}, two coefficients of variation are estimated:
\itemize{
\item the CV before the replication stage (see \code{pre_replicate_cv_prior_mu})
\item the CV after the replication stage (see \code{cv_prior_mu})

The meaning of these CV estimates depends on the type of replicates. If the
replicates are biological replicates (i.e. independently processed), then
\code{cv} estimates the noise in the preprocessing before the PCR, and
\code{pre_replicate_cv} estimates the noise from anything before preprocessing
(e.g. sampling noise and all other unexplained variation). In contrast, if
the replicates are technical replicates (i.e. several PCR runs of the same
preprocessed sample), then \code{cv} estimates only unexplained PCR noise
(should be close to zero), and \code{pre_replicate_cv} estimates all other noise
(including preprocessing noise.)
}

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage (\code{pre_replicate_cv}):
\verb{Truncated normal}
}
}
\seealso{
\link{LOD_estimate_dPCR} for a limit of detection model specialised for
dPCR.

\link{noise_estimate_dPCR_params} for jointly estimating uncertain
parameters of the assay.

Other noise models: 
\code{\link{noise_estimate}()},
\code{\link{noise_estimate_constant_var}()},
\code{\link{noise_estimate_dPCR_params}()}
}
\concept{noise models}
