% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_dPCR_params}
\alias{noise_estimate_dPCR_params}
\title{Estimate measurement noise for digital PCR data with priors for assay
parameters}
\usage{
noise_estimate_dPCR_params(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  total_partitions_observe = FALSE,
  max_partitions_prior_lower = 20000,
  max_partitions_prior_upper = 40000,
  partition_loss_mean_prior_lower = 0.05,
  partition_loss_mean_prior_upper = 0.3,
  partition_loss_variation_prior_lower = 0.5,
  partition_loss_variation_prior_upper = 2,
  partition_loss_max = 0.5,
  volume_scaled_prior_lower = 1e-06,
  volume_scaled_prior_upper = 1e-04,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  prePCR_noise_type = "log-normal",
  use_taylor_approx = TRUE
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of analyzed
concentrations (pre-PCR noise). Note that this does \emph{not} include the
technical noise of the digital PCR. This is because the dPCR noise is
explicitly modeled (using the total number of partitions and conversion
factor). Moreover, when \code{replicates=TRUE}, noise before the replication
stage is separately modeled, so \code{cv_prior_mu} only describes any
remaining pre-PCR noise (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of analyzed concentrations.}

\item{total_partitions_observe}{If TRUE, the total number of partitions is
taken from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_partitions_col} is specified in \code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{max_partitions_prior_lower}{Prior (lower bound) for the maximum total
number of dPCR partitions. This is usually defined by the manufacturer of
the dPCR system/chip used, which supports a certain maximum number of
partitions. If you know the exact dPCR system and its maximum partition
number, you can set both \code{max_partitions_prior_lower} and
\code{max_partitions_prior_upper} to this value. Otherwise, this prior can be
used to set a broad lower and upper bound for the maximum number of
partitions, to reflect a range of popular dPCR systems/chips.}

\item{max_partitions_prior_upper}{Prior (upper bound) for the maximum total
number of dPCR partitions (see \code{max_partitions_prior_lower} for details.)}

\item{partition_loss_mean_prior_lower}{Prior (5\% quantile) for the mean
relative partition loss. A certain proportion of partitions in a dPCR run
is typically invalid and discarded from the concentration estimate. This
prior can be used to set a lower and upper bound for the mean proportion of
partitions lost. Note that for proportions close to 0 or to
\code{partition_loss_max} (see below), the resulting mean partition loss can
differ from what is specified here, because we internally
translate this prior to the logit scale.}

\item{partition_loss_mean_prior_upper}{Prior (95\% quantile) for the mean
relative partition loss (see \code{partition_loss_mean_prior_lower} for
details). In well-functioning dPCR assays, the mean proportion of lost
partitions should not be very high (definitely below 50\%).}

\item{partition_loss_variation_prior_lower}{Prior (5\% quantile) for the
variation in the number of invalid partitions across dPCR runs. The
proportion of partitions lost typically varies between dPCR runs. We thus
model this proportion as logit-normal distributed, with mean
(approximately) defined by \code{partition_loss_mean_prior} and logit-level
standard deviation \code{sigma}. You can use
\code{partition_loss_variation_prior_lower} and
\code{partition_loss_variation_prior_upper} to set a lower and upper bound for
\code{sigma}.}

\item{partition_loss_variation_prior_upper}{Prior (95\% quantile) for the
variation in the number of invalid partitions across dPCR runs (see
\code{partition_loss_variation_prior_lower} for details).}

\item{partition_loss_max}{The maximum proportion of partitions that can be
lost in a valid dPCR run. During quality control, runs where the proportion
of invalid partitions is above some threshold (e.g. 50\%) are often
discarded. This parameter can be used to represent such a QC threshold.}

\item{volume_scaled_prior_lower}{Prior (lower bound) on the conversion factor
(partition volume scaled by the dilution of wastewater in the assay) for
the dPCR reaction. See details for further explanation.}

\item{volume_scaled_prior_upper}{Prior (upper bound) on the conversion factor
(partition volume scaled by the dilution of wastewater in the assay) for
the dPCR reaction. If this is identical to \code{volume_scaled_prior_lower}, the
conversion factor will be fixed and not estimated.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} is used.}

\item{use_taylor_approx}{If TRUE (default), a Taylor expansion approximation
is used to estimate the CV of measurements under pre-PCR noise. The
approximation is very accurate, unless concentrations are extremely high
(so high that the quality of the measurements from dPCR would anyway be
questionable).}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option models concentration measurements using an error
model specialized for digital PCR (e.g. digital droplet PCR). This option
also jointly estimates the uncertain parameters of the dPCR assay. For a
faster approximation assuming fixed assay parameters, see
\code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}}.

This model predicts more relative variation at smaller
concentrations, which often leads to a better model fit, even for
measurements from other quantification methods such as qPCR.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
The conversion factor (see \code{volume_scaled_prior_lower},
\code{volume_scaled_prior_upper}) is the partition volume v multiplied with a
scaling factor s. The scaling factor accounts for concentration differences
between the sample and the reaction mix, for example due to extraction or
adding of reagents. For example, if the partition volume is 4.5e-7 mL and
the scaling factor is 100:3 (i.e. 100 gc/mL in the original sample
correspond to 3 gc/mL in the PCR reaction), then the overall conversion
factor is 4.5e-7 * 100 / 3 = 1.5e-5.

When \code{replicates=TRUE}, two coefficients of variation are estimated:
\itemize{
\item the CV before the replication stage (see \code{pre_replicate_cv_prior_mu})
\item the CV after the replication stage (see \code{cv_prior_mu})

The meaning of these CV estimates depends on the type of replicates. If the
replicates are biological replicates (i.e. independently processed), then
\code{cv} estimates the noise in the preprocessing before the PCR, and
\code{pre_replicate_cv} estimates the noise from anything before preprocessing
(e.g. sampling noise and all other unexplained variation). In contrast, if
the replicates are technical replicates (i.e. several PCR runs of the same
preprocessed sample), then \code{cv} estimates only unexplained PCR noise
(should be close to zero), and \code{pre_replicate_cv} estimates all other noise
(including preprocessing noise.)
}

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
\item maximum number of total partitions in dPCR: \code{Uniform}
\item mean proportion of lost partitions in dPCR: \code{Normal (logit-level)}
\item variation of proportion of lost partitions: \verb{Truncated normal (logit-level)}
\item conversion factor for dPCR: \code{Uniform}
\item coefficient of variation of concentration before the replication stage (\code{pre_replicate_cv}):
\verb{Truncated normal}
}
}
\seealso{
\link{LOD_estimate_dPCR} for a limit of detection model specialised for
dPCR.

\link{noise_estimate_dPCR} for a faster dPCR noise model with fixed assay
parameters.

Other noise models: 
\code{\link{noise_estimate}()},
\code{\link{noise_estimate_constant_var}()},
\code{\link{noise_estimate_dPCR}()}
}
\concept{noise models}
