% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{outliers_none}
\alias{outliers_none}
\title{Do not model outliers}
\usage{
outliers_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option does not model outliers in sampled concentrations.
}
\seealso{
Other outlier models: 
\code{\link{outliers_estimate}()}
}
\concept{outlier models}
