% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{place_knots}
\alias{place_knots}
\title{Places knots for fitting B-splines to a time series}
\usage{
place_knots(
  length_R,
  forecast_horizon,
  knot_distance,
  partial_window,
  partial_generation,
  fix_forecast = FALSE
)
}
\arguments{
\item{length_R}{Length of the Rt time series}

\item{forecast_horizon}{Number of days to forecast}

\item{knot_distance}{Normal distance between knots}

\item{partial_window}{Window with only partial data towards the present in
which the knot distances will be shorter. This is to avoid erroneous
extrapolation of splines towards the present.}

\item{partial_generation}{A certain quantile of the generation time
distribution. The first "fully informed" knot is placed at that distance
after the partial window. The reason for this is that while the number of
infections is already sufficiently informed after the partial window, the
trend in infections is typically very volatile, so we place the first
variable knot after approximately one generation.}
}
\value{
A vector with knot positions.
}
\description{
Places knots for fitting B-splines to a time series
}
\details{
The splines are forced to a zero slope towards the present and for
forecasting.
}
\keyword{internal}
