% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Title}
\usage{
plot_concentration(
  results,
  measurements = NULL,
  include_noise = TRUE,
  median = FALSE
)
}
\arguments{
\item{include_noise}{}
}
\value{

}
\description{
Title
}
