% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_growth_report}
\alias{plot_growth_report}
\title{Plot a growth report}
\usage{
plot_growth_report(result, date = NULL, partial_prob = 0.8)
}
\arguments{
\item{result}{Results object returned by \code{\link[=EpiSewer]{EpiSewer()}} after model fitting. In
contrast to other plotting functions, this cannot be a list of multiple
result objects, because growth report plots are always for a single model
fit.}

\item{date}{The date for which the growth report should be plotted. If NULL,
the most recent date for which a reliable report can be provided is
automatically selected, see \code{partial_prob}.}

\item{partial_prob}{To select the most recent reliable date, we subtract a
certain quantile of the shedding load distribution from the current date.
For example, if \code{partial_prob=0.8} (default), we select the date for which
80\% of the shedding load of individuals infected before this date has been
shed.}
}
\value{
A growth report plot showing the probability that infections have
been growing for at least 3, 7, 14, 21, and 28 days, respectively. Can be
further manipulated using \link{ggplot2} functions to adjust themes and scales,
and to add further geoms.
}
\description{
For a given date, the growth report shows the probability that
infections have been growing for at least 3, 7, 14, 21, and 28 days,
respectively. The report uses a diverging bar plot which is scaled between
"very unlikely" (0\% posterior probability) and "very likely" (100\%
posterior probability).
}
