% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_LOD}
\alias{plot_LOD}
\title{Plot limit of detection}
\usage{
plot_LOD(modeldata)
}
\arguments{
\item{modeldata}{A \code{modeldata} object as returned by \code{\link[=LOD_assume]{LOD_assume()}}.}
}
\value{
A plot showing the probability of non-detection (i.e. zero
measurement) for different concentrations below and above the assumed LOD.
Can be further manipulated using \link{ggplot2} functions to adjust themes and
scales, and to add further geoms.
}
\description{
Helper function to visualize the assumed limit of detection.
}
\details{
This function can also be applied to \code{modeldata} objects which have
been passed through several other modeling functions, as long as
\code{\link[=LOD_assume]{LOD_assume()}} was applied at some point.
}
\examples{
# Particles arrive within the same day
residence_dist_assume(residence_dist = c(1))

# Particles always arrive after one day
residence_dist_assume(residence_dist = c(0, 1))

# 1/4 of particles only arrives after one day
residence_dist_assume(residence_dist = c(0.75, 0.25))
}
