% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_prior_partitions}
\alias{plot_prior_partitions}
\title{Plot prior distributions for total number of partitions}
\usage{
plot_prior_partitions(modeldata, n_draws = 1000, show_draws = 50, seed = 0)
}
\arguments{
\item{modeldata}{A model component as returned by
\code{\link[=noise_estimate_dPCR_params]{noise_estimate_dPCR_params()}}, or an \code{EpiSewer} job/result object.}

\item{n_draws}{Number of draws to simulate from the prior distributions.}

\item{show_draws}{Number of example draws to show in the plots.}

\item{seed}{Seed for random number generation to ensure reproducibility.}
}
\value{
A grid of plots showing the distribution of partition loss, 95\%
intervals for valid partitions, and distributions of mean and standard
deviation of valid partitions.
}
\description{
Plot prior distributions for total number of partitions
}
\keyword{internal}
