% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{Visually compare prior and posterior of a model parameter}
\usage{
plot_prior_posterior(result, param_name)
}
\arguments{
\item{result}{Results object returned by \code{\link[=EpiSewer]{EpiSewer()}} after model fitting. In
contrast to other plotting functions, this cannot be a list of multiple
result objects, because prior-posterior plots of multiple results
simultaneously are currently not supported.}

\item{param_name}{Name of the single parameter to be plotted. This can either
be the raw name of the parameter in the stan model, or it can be the
semantic name from the single parameter dictionary (see details). Note that
this must be a single parameter, it cannot be a parameter array. Also, only
original parameters (not transformed parameters) for which a prior can be
specified in \code{EpiSewer} are supported.}
}
\value{
A plot showing the density of the prior (grey) and posterior (blue)
for the respective parameter. Can be further manipulated using \link{ggplot2}
functions to adjust themes and scales, and to add further geoms.
}
\description{
Visually compare prior and posterior of a model parameter
}
\details{
The following parameters can be visualized (if in the model):
\itemize{
\item \code{measurement_noise_cv} (nu_upsilon_a): Coefficient of variation (measurement noise)
\item \code{ddPCR_droplets} (nu_upsilon_b): Number of droplets in ddPCR
\item \code{ddPCR_scaling} (nu_upsilon_c): Concentration conversion factor in ddPCR
\item \code{pre_replicate_cv} (nu_psi): Coefficient of variation (pre-PCR noise)
\item \code{load_variation_cv} (nu_zeta): Individual-level coefficient of load variation
\item \code{infection_overdispersion} (I_xi): Overdispersion of infections
\item \code{seeding_intercept} (iota_log_seed_intercept): Initial number of infections
}
}
