% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelcomponent.R
\name{print.EpiSewer_module}
\alias{print.EpiSewer_module}
\title{Print an EpiSewer module}
\usage{
\method{print}{EpiSewer_module}(x, ...)
}
\description{
Print an EpiSewer module
}
\keyword{internal}
