% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelcomponent.R
\name{print.model_component}
\alias{print.model_component}
\title{Print a model_component}
\usage{
\method{print}{model_component}(x, ...)
}
\description{
Print a model_component
}
\keyword{internal}
