% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{print.modeldata}
\alias{print.modeldata}
\title{Print a \code{modeldata} object}
\usage{
\method{print}{modeldata}(x, type = "structure")
}
\arguments{
\item{type}{If "structure", the model structure is printed. If "data", the
modeldata content (data, inits, metainfo) is printed.}
}
\description{
Print a \code{modeldata} object
}
\keyword{internal}
