% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{qexpgamma}
\alias{qexpgamma}
\title{Exponential-Gamma distribution}
\usage{
qexpgamma(p, shape, scale)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{shape}{Shape of the Exponential-Gamma distribution.}

\item{scale}{Scale of the Exponential-Gamma distribution.}
}
\description{
The Exponential-Gamma (EG) distribution is also known as the Lomax
distribution and has a characteristic long tail.
}
