% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_approx}
\alias{R_estimate_approx}
\title{Estimate Rt using an approximation of the generative renewal model}
\usage{
R_estimate_approx(
  inf_sd_prior_mu = 0.05,
  inf_sd_prior_sigma = 0.025,
  inf_smooth = 0.75,
  inf_trend_smooth = 0.75,
  inf_trend_dampen = 0.9,
  knot_distance = 7,
  spline_degree = 3,
  R_window = 1
)
}
\arguments{
\item{inf_sd_prior_mu}{Prior (mean) on the standard deviation of daily
changes in infections. Infection numbers are modeled on the log scale, thus
the standard deviation can be roughly interpreted in terms of percentage
changes. A higher standard deviation will lead to more volatile infections
and more uncertainty in R estimates. The default (\code{inf_sd_prior_mu = 0.05})
allows daily growth/decline rates of up +-10\%. We suggest to only decrease
this if you are very certain that the modeled scenario has lower maximum
growth/decline rates, and to increase this if you expect a more extreme
scenario. Note that the smoothness of the infection curve is also
influenced by the exponential smoothing parameters (\code{inf_smooth},
\code{inf_trend_smooth}). Still, \code{inf_sd_prior_mu} is the primary hyperparameter
to adjust if you think the estimated infection trajectory is too volatile /
not flexible enough.}

\item{inf_sd_prior_sigma}{Prior (standard deviation) on standard deviation of
daily changes of infections. This reflects uncertainty around
\code{inf_sd_prior_mu}. The default (\code{inf_sd_prior_mu=0.025}) allows the
standard deviation to be roughly 0.5 higher or lower than
\code{inf_sd_prior_mu}. For the default setting with \code{inf_sd_prior_mu = 0.05},
this means that growth/decline rates of up to +-20\% are still supported by
the prior.}

\item{inf_smooth}{Smoothing parameter (alpha) for infections. The infection
time series is smoothed based on earlier time steps, with exponentially
decaying weights, as controlled by the alpha parameter. The default is 0.5,
\emph{smaller} values will lead to \emph{stronger} smoothing, and \emph{larger} values to
\emph{less} smoothing.}

\item{inf_trend_smooth}{Trend smoothing parameter (beta) for infections. The
exponential trend in infections is also smoothed based on the trend over
earlier time steps, with exponentially decaying weights, as controlled by
the beta parameter. Default is 0.5, \emph{smaller} values will lead to
\emph{stronger} smoothing of the trend (useful for longer generation times),
and \emph{larger} values to \emph{less} smoothing of the trend (useful for shorter
generation times).}

\item{inf_trend_dampen}{Trend damping parameter (phi) for infections. The
exponential trend in infections is dampened over time (exponential growth
will not continue for ever). This mainly influences the trend towards the
present. For the default value (0.9), the trend will only roughly half as
strong after two weeks. Note that increasing phi to close to 1 can reduce
sampling speed.}

\item{knot_distance}{The infection time series is smoothed using penalized
B-splines. The distance between spline breakpoints (knots) is 7 days by
default. Placing knots further apart increases the smoothness of the
infection time series and can speed up model fitting. Placing knots closer
to each other increases the volatility of the infections. Note however that
the uncertainty of R estimates produced by \code{\link[=R_estimate_approx]{R_estimate_approx()}} is also
sensitive to the smoothness of the infection time series, i.e. changing the
knot distance can make the R estimates appear more or less uncertain. This
effect is however rather small since the regularization of the spline
coefficients adjusts for the knot distance.}

\item{spline_degree}{Degree of the spline polynomials (default is 3 for cubic
splines).}

\item{R_window}{Smoothing window size for R estimates. Default is 1 (i.e. no
smoothing). This is provided for compatibility with the EpiEstim method by
Cori et al., which assumes that R stays constant over a certain time
window. However, this can artificially reduce the uncertainty of the R
estimates and is therefore generally not recommended. It also means that R
cannot be estimated up to the present because the smoothing window must be
centered.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=R_estimate_gp]{R_estimate_gp()}} or other
fully generative methods instead.

This option estimates the effective reproduction number Rt using
an approximation of the generative renewal model. It can be considerably
faster than the fully generative options, but is often less exact. It is
recommended to check the results of this method against a fully generative
version like \code{\link[=R_estimate_gp]{R_estimate_gp()}} before using it for real-time R
estimation. See the details for an explanation of the method and its
limitations.
}
\details{
This method uses an approximation of the renewal model to estimate
the effective reproduction number. Instead of generating the infections
through a renewal process, the infection time series is estimated using a
non-parametric smoothing prior that mimics the characteristics of a renewal
process. Rt is then computed from the infection time series by applying the
classical renewal equation. This still gives Bayesian estimates of Rt and
the infection time series, but is less exact than a fully generative model.

To smooth the infection time series, penalized B-splines similar to
the ones in \code{\link[=R_estimate_splines]{R_estimate_splines()}} are used, but they are applied to the
expected number of infections on the log scale, not Rt. The spline
coefficients are themselves smoothed using an exponential smoothing prior
with a trend component. The trend component is important as it reflects the
default assumption of constant transmission dynamics that leads to
exponential growth in infections. This ensures consistent results of
\code{R_estimate_approx()} also towards the present.

The method can also model infection noise as specified by
\code{\link[=infection_noise_estimate]{infection_noise_estimate()}}. To account for the autocorrelation of
infection noise, a correction that applies the renewal model specifically
to the noise component is used. This correction is most accurate when R is
close to 1 and less accurate when it is far from 1.

The main limitation of \code{R_estimate_approx()} is that its priors and
hyperparameters are less interpretable and thus difficult to specify. In
particular the assumed generation time distribution is only used for R
estimation but does not inform the smoothness of the expected infection
time series. Instead, the smoothness is determined by the spline settings
and exponential smoothing parameters, which are difficult to translate into
a generation time assumption. The best approach is therefore to compare
estimates from \code{R_estimate_approx()} with those from another method like
\code{\link[=R_estimate_splines]{R_estimate_splines()}} for a pathogen of interest and to carefully adjust
the smoothing hyperparameters if needed.
}
\seealso{
Other Rt models: 
\code{\link{R_estimate_changepoint_splines}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_gp}()},
\code{\link{R_estimate_piecewise}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_smooth_derivative}()},
\code{\link{R_estimate_splines}()}
}
\concept{Rt models}
