% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{read_scalar_params}
\alias{read_scalar_params}
\title{Read scalar parameters from a text file}
\usage{
read_scalar_params(model)
}
\arguments{
\item{model}{The model specification, used to identify the correct text file.}
}
\value{
A character vector with the names of the scalar parameters.
}
\description{
Read scalar parameters from a text file
}
\keyword{internal}
