% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{read_scalar_vars}
\alias{read_scalar_vars}
\title{Read scalar variables from a text file}
\usage{
read_scalar_vars(model)
}
\arguments{
\item{model}{The model specification, used to identify the correct text file.}
}
\value{
A character vector with the names of the scalar variables.
}
\description{
Read scalar variables from a text file
}
\keyword{internal}
