% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{rep_each_v}
\alias{rep_each_v}
\title{Vectorized version of \code{rep} with \code{each} argument, \code{each} can be a vector}
\usage{
rep_each_v(x, each)
}
\description{
Vectorized version of \code{rep} with \code{each} argument, \code{each} can be a vector
}
\keyword{internal}
