% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy.R
\name{resolve}
\alias{resolve}
\title{Resolve a lazy EpiSewer object}
\usage{
resolve(x, ...)
}
\arguments{
\item{x}{A lazy object of class \code{tbp}, \code{tbe}, or \code{tbc}.}

\item{...}{Further arguments passed to the method.}
}
\value{
The resolved value or updated modeldata.
}
\description{
Resolves a lazy \code{tbp}, \code{tbe}, or \code{tbc} object by evaluating it
once all required inputs are available.
}
\keyword{internal}
