% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{sample_effects_estimate_weekday}
\alias{sample_effects_estimate_weekday}
\title{Estimate weekday sample effects}
\usage{
sample_effects_estimate_weekday(effect_prior_mu = 0, effect_prior_sigma = 1)
}
\arguments{
\item{effect_prior_mu}{Prior (mean) on the regression coefficients.}

\item{effect_prior_sigma}{Prior (standard deviation) on the regression
coefficients.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option uses a log-linear regression model to estimate
sample weekday effects on the concentration. Concentrations can be
influenced by the time between sampling and shipping to the lab
(age-of-sample effect), and if shipment follows a weekly batch scheme, the
sampling weekday is a good proxy for the age at shipment.
}
\details{
Effects are estimated for weekdays Monday - Saturday, with Sunday as
the baseline. \code{EpiSewer} will fit a fixed-effects log-linear model, random
effects are currently not supported.

The priors of this component have the following functional form:
\itemize{
\item regression coefficients: \code{Normal}
}
}
\seealso{
Other sample effect models: 
\code{\link{sample_effects_estimate_matrix}()},
\code{\link{sample_effects_none}()}
}
\concept{sample effect models}
