% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{sample_effects_matrix_}
\alias{sample_effects_matrix_}
\title{Add a sample effects design matrix to modeldata (shared internal)}
\usage{
sample_effects_matrix_(
  design_matrix,
  effect_prior_mu,
  effect_prior_sigma,
  modeldata
)
}
\description{
Used by \code{\link[=sample_effects_estimate_matrix]{sample_effects_estimate_matrix()}} and
\code{\link[=sample_effects_estimate_weekday]{sample_effects_estimate_weekday()}}.
}
\keyword{internal}
