% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{sample_effects_none}
\alias{sample_effects_none}
\title{Do not model sample effects}
\usage{
sample_effects_none()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option does not model effects of sample covariates on the
concentrations.
}
\seealso{
Other sample effect models: 
\code{\link{sample_effects_estimate_matrix}()},
\code{\link{sample_effects_estimate_weekday}()}
}
\concept{sample effect models}
