% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{sampler_stan_mcmc}
\alias{sampler_stan_mcmc}
\title{Title}
\usage{
sampler_stan_mcmc(
  chains = 4,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt_delta = 0.99,
  max_treedepth = 15,
  step_size = 0.1,
  parallel_chains = NULL,
  threads_per_chain = 1,
  seed = 0,
  refresh = 200,
  show_messages = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
