% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{seeding_estimate_growth}
\alias{seeding_estimate_growth}
\title{Estimate seeding infections with a time-varying growth rate}
\usage{
seeding_estimate_growth(
  intercept_prior_q5 = NULL,
  intercept_prior_q95 = NULL,
  growth_change_prior_mu = 0,
  growth_change_prior_sigma = 0.01,
  extend = TRUE
)
}
\arguments{
\item{intercept_prior_q5}{Prior (5\% quantile) on the initial number of
infections. Can be interpreted as an approximate lower bound. If NULL
(default), this is computed from a crude empirical estimate of the number
of cases (see details).}

\item{intercept_prior_q95}{Prior (95\% quantile) on the initial number of
infections. Can be interpreted as an approximate upper bound. If NULL
(default), this is computed from a crude empirical estimate of the number
of cases (see details).}

\item{growth_change_prior_mu}{Prior (mean) on the daily standard deviation of
the random walk for the epidemic growth rate.}

\item{growth_change_prior_sigma}{Prior (standard deviation) on the daily
standard deviation of the random walk for the epidemic growth rate.}

\item{extend}{Should the seeding phase be extended when concentrations are
very low at the start of the measurement time series? If \code{TRUE}, then the
seeding phase will be extended to the first date with three consecutive
detects (i.e. non-zero measurements). Before that date, the reproduction
number will be retrospectively computed based on the seeded infections.
Explicit modeling of the Rt time series will only begin after the seeding
phase. This option avoids sampling problems when estimating Rt from very
low infection numbers during a period with many non-detects. Note that
estimated reproduction numbers are not necessarily meaningful during
periods with very low infection numbers, as transmission dynamics may be
dominated by chance events and importations.}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option estimates an exponential growth of infections at the
start of the modeled time period, with the growth rate varying over time.
}
\details{
The seeding phase has the length of the maximum generation time
(during this time, the renewal model cannot be applied). It is here assumed
that the expected number of new infections follows an exponential growth or
decline process during this time period. The exponential growth rate of the
seeding phase follows a random walk that ends with a growth rate
representing the first estimated reproduction number at the start of the
modeling phase. This means that the intercept of the seeding phase growth
rate depends on the \code{R_start_prior} provided in the \verb{R_estimate_*}
component.

If the lower and upper intervals of the prior for the initial number
of infections, i.e. \code{intercept_prior_q5} or \code{intercept_prior_q95} are not
specified by the user, \code{EpiSewer} will compute a rough median empirical
estimate of the number of cases using the supplied wastewater measurements
and shedding assumptions, and then infer the missing quantiles based on
this. If none of the quantiles are provided, they are set to be roughly
1/10 and 10 times the empirical median estimate. We note that this is a
violation of Bayesian principles (data must not be used to inform priors) -
but a neglectable one, since it only ensures that the seeding is modeled on
the right order of magnitude and does not have relevant impacts on later Rt
estimates.

The priors of this component have the following functional form:
\itemize{
\item initial number of infections (log scale): \code{Normal}
\item standard deviation of the random walk on the growth rate: \verb{Truncated normal}

The priors for these parameters are determined based on the user-supplied
arguments, using appropriate transformations and the two-sigma-rule of
thumb.
}

Credits to Samuel Brand and the authors of the EpiAware toolkit for
the idea to back-calculate the growth rate of the seeding phase from the
initial reproduction number.
}
