% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{set_fit_opts}
\alias{set_fit_opts}
\title{Configure the model fitting}
\usage{
set_fit_opts(sampler = sampler_stan_mcmc(), model = model_stan_opts())
}
\arguments{
\item{sampler}{Which type of sampler should be used for model fitting?
Currently supported are
\itemize{
\item \code{\link[=sampler_stan_mcmc]{sampler_stan_mcmc()}}: default, recommended for inference
\item \code{\link[=sampler_stan_pathfinder]{sampler_stan_pathfinder()}}: potentially inaccurate, recommended for
preview only
}}

\item{model}{Details about the model file to be used, see
\code{\link[=model_stan_opts]{model_stan_opts()}}. This also makes it possible for users to supply
customized models.}
}
\value{
A \code{list} with the model fitting options.
}
\description{
Sets model fitting options in \code{EpiSewer}.
}
\examples{
ww_data <- ww_data_SARS_CoV_2_Zurich
ww_assumptions <- ww_assumptions_SARS_CoV_2_Zurich

ww_fit_opts <- set_fit_opts(
  sampler = sampler_stan_mcmc(
    chains = 2,
    parallel_chains = 2,
    iter_warmup = 400,
    iter_sampling = 400,
    seed = 42 # ensures reproducibility
  )
)

\dontrun{ww_result <- EpiSewer(
  data = ww_data,
  assumptions = ww_assumptions,
  fit_opts = ww_fit_opts
)}
}
