% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{set_prior_trunc_normal}
\alias{set_prior_trunc_normal}
\title{Define a truncated normal prior in modeldata}
\usage{
set_prior_trunc_normal(
  param,
  mu = NULL,
  sigma = NULL,
  two_sigma = NULL,
  q5 = NULL,
  q95 = NULL
)
}
\arguments{
\item{param}{Name of the parameter for which the prior is defined.}

\item{mu}{Mean (not accounting for truncation).}

\item{sigma}{Standard deviation (not accounting for truncation).}

\item{two_sigma}{Two times the standard deviation (not accounting for
truncation). Useful for defining priors via the two-sigma rule-of-thumb
(approximately 95\% of probability mass).}

\item{q5}{Lower quantile (5\%).}

\item{q95}{Upper quantile (95\%).}
}
\value{
Prior specification for modeldata.
}
\description{
This function defines a truncated normal prior (truncation at
zero) for a parameter in modeldata.
}
\keyword{internal}
