% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{sewer_compile}
\alias{sewer_compile}
\title{Compile EpiSewer models}
\usage{
sewer_compile(model = NULL, force_recompile = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{A character vector with a specific model to compile. If NULL
(default), all models are compiled.}

\item{force_recompile}{If TRUE, then models will be recompiled even if they
have already been successfully compiled.}

\item{verbose}{If TRUE, warnings and detailed errors from the compilation are
printed. This can help to diagnose compilation issues.}
}
\description{
The stan models used by EpiSewer need to be compiled for your
device. This is only necessary once, after installing or updating the
package. This function compiles all models from the package.
}
\details{
If one or several models are not successfully compiled, please
ensure that \code{cmdstan} is properly set up and try updating it to a newer
version using \code{\link[cmdstanr:install_cmdstan]{cmdstanr::install_cmdstan()}}. If the problem persists,
please run \code{sewer_compile(verbose = TRUE)} and post the output in
a new issue on GitHub, along with your \code{\link[cmdstanr:set_cmdstan_path]{cmdstanr::cmdstan_version()}}.
}
