% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{sewer_data}
\alias{sewer_data}
\title{Specify observation data}
\usage{
sewer_data(measurements = NULL, flows = NULL, cases = NULL, ...)
}
\arguments{
\item{measurements}{A \code{data.frame} with measured concentrations of the
pathogen of interest. Will be automatically passed to
\code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{flows}{A \code{data.frame} with wastewater flow volumes at the sampling site
for each day. Will be automatically passed to \code{\link[=flows_observe]{flows_observe()}}.}

\item{cases}{A \code{data.frame} of case numbers with each row representing one
day. Must have at least a column with dates and a column with case numbers.
This data is not used for model fitting, but for calibration of the
\code{load_per_case} assumption. Will be automatically passed to
\code{\link[=load_per_case_calibrate]{load_per_case_calibrate()}}.}

\item{...}{Further observations to be passed to \code{\link[=EpiSewer]{EpiSewer()}}.}
}
\value{
A \code{list} with all observations supplied. Can be passed to the \code{data}
argument in \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Specify wastewater observation data such as concentration
measurements and flows. This is a convenience function to collect all
observation data in one object.
}
