% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{sewer_pull_docker}
\alias{sewer_pull_docker}
\title{Pull the latest EpiSewer docker image}
\usage{
sewer_pull_docker()
}
\description{
This function pulls the latest EpiSewer docker image from the
github container registry. This is necessary to use the docker backend for
model fitting.
}
