% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{shedding_dist_assume}
\alias{shedding_dist_assume}
\title{Assume a shedding load distribution}
\usage{
shedding_dist_assume(shedding_dist = NULL, shedding_reference = NULL)
}
\arguments{
\item{shedding_dist}{A numeric vector representing a discrete shedding load
distribution, with elements describing the share of load shed 0 days, 1
day, 2 days, and so on after the start of shedding.}

\item{shedding_reference}{Is the shedding load distribution relative to the
day of \code{"infection"} or the day of \code{"symptom_onset"}? This is important because
shedding load distributions provided in the literature are sometimes by
days since infection and sometimes by days since symptom onset. If
\code{shedding_reference="symptom_onset"}, EpiSewer also needs information about
the incubation period distribution (see \code{\link[=incubation_dist_assume]{incubation_dist_assume()}}).}
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This option assumes a fixed shedding load distribution for the
shedding model in \code{EpiSewer}.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
