% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{shedding_dist_estimate_mixture}
\alias{shedding_dist_estimate_mixture}
\title{Estimate a mixture of several shedding load distributions}
\usage{
shedding_dist_estimate_mixture(
  shedding_dist_list,
  prior_weights = NULL,
  shedding_reference = NULL,
  weight_alpha = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{shedding_dist_list}{A list of shedding load distributions. Each entry
must be a numeric vector representing a discrete shedding load
distribution, with elements describing the share of load shed 0 days, 1
day, 2 days, and so on after the start of shedding.}

\item{prior_weights}{A numeric vector of the same length as
\code{shedding_dist_list} with prior weights for the different shedding load
distributions. If \code{NULL}, the distributions are assumed to be equally
likely (symmetric Dirichlet prior). Will be normalized to sum to the number
of distributions.}

\item{shedding_reference}{Is the shedding load distribution relative to the
day of \code{"infection"} or the day of \code{"symptom_onset"}? This is important
because shedding load distributions provided in the literature are
sometimes by days since infection and sometimes by days since symptom
onset. If \code{shedding_reference="symptom_onset"}, EpiSewer also needs
information about the incubation period distribution (see
\code{\link[=incubation_dist_assume]{incubation_dist_assume()}}).}

\item{weight_alpha}{Concentration parameter of the Dirichlet prior for the
mixture probabilities. The default is \code{weight_alpha=1}, i.e. a uniform
distribution over all combinations of weights. This will sample various
mixtures of the distributions provided. If on the other hand a very small
\code{weight_alpha} is chosen, the distributions from \code{shedding_dist_list} are
rather considered separately, i.e. in each posterior sample, one of the
distributions is given almost all the weight. Note however that for very
small alpha values the Dirichlet prior becomes difficult to sample with
MCMC, so it may be necessary to increase \code{adapt_delta} to avoid divergent
transitions (see \code{\link[=sampler_stan_mcmc]{sampler_stan_mcmc()}}).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option allows to model a mixture of several shedding load
distributions with a Dirichlet prior over the mixture weights.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}

Other {shedding load distribution functions}: 
\code{\link{load_per_case_assume}()}
}
\concept{{shedding load distribution functions}}
