% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{slot_module_map}
\alias{slot_module_map}
\title{Map component slots to their modules (for the model structure summary)}
\usage{
slot_module_map()
}
\description{
Map component slots to their modules (for the model structure summary)
}
\keyword{internal}
