% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{standata_update_metainfo}
\alias{standata_update_metainfo}
\title{Update meta information in standata based on available variables}
\usage{
standata_update_metainfo(standata)
}
\description{
This update function is for all meta information that depends on standata
from several functions. There are also functions which add meta information
directly, in particular when this meta information cannot be solely inferred
from standata. This function is designed such that calling it will never do
harm to the standata object and not throw errors if something is missing in
the standata object.
}
