% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{template_model_helpers}
\alias{template_model_helpers}
\title{Template function}
\usage{
template_model_helpers()
}
\value{
A \code{model_component} object describing this part of the model
specification. Pass it to the corresponding module function (e.g.
\code{\link[=model_measurements]{model_measurements()}}, \code{\link[=model_infections]{model_infections()}}), which in turn is supplied
to \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
This function does not actually do anything. It only serves as
a template for the documentation of other functions using inheritance.
}
\keyword{internal}
