% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{test_run}
\alias{test_run}
\title{Run a quick model fitting test with minimal iterations.}
\usage{
test_run(job)
}
\arguments{
\item{job}{An \code{EpiSewerJob} or \code{EpiSewerJobResult} object.}
}
\value{
An \code{EpiSewerJobResult} object.
}
\description{
Modifies sampler settings to use very few iterations and then
calls \code{\link[=run]{run()}}. Useful for quickly checking that a model specification is
valid and that the model can be fitted without errors.
}
\keyword{internal}
