% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_compile.R
\name{trim_spec}
\alias{trim_spec}
\title{Trim large objects from the spec tree after compilation}
\usage{
trim_spec(spec)
}
\description{
Trim large objects from the spec tree after compilation
}
\keyword{internal}
