% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{trunc_normal_mean}
\alias{trunc_normal_mean}
\title{Calculate the mean of a truncated normal distribution (truncated below zero)
with mean mu and standard deviation sigma.}
\usage{
trunc_normal_mean(mu, sigma)
}
\arguments{
\item{mu}{Mean of the untruncated normal distribution}

\item{sigma}{Standard deviation of the untruncated normal distribution}
}
\value{
The mean of the truncated normal distribution
}
\description{
Calculate the mean of a truncated normal distribution (truncated below zero)
with mean mu and standard deviation sigma.
}
