% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy.R
\name{update_env}
\alias{update_env}
\title{Add required data and assumptions to an environment}
\usage{
update_env(
  f_env = list(),
  data = list(),
  assumptions = list(),
  required_data = c(),
  required_assumptions = c()
)
}
\arguments{
\item{f_env}{The environment to be updated}

\item{data}{A list with data}

\item{assumptions}{A list with assumptions}

\item{required_data}{A character vector with the names of required data.
Different options for one item can be provided using "|" as a separator.
EpiSewer will then check if at least one of the options is present.}

\item{required_assumptions}{A character vector with the names of required
assumptions. Different options for one item can be provided using "|" as a
separator. EpiSewer will then check if at least one of the options is
present.}
}
\value{
The updated environment, where all required data and assumptions that
were provided in \code{data} and \code{assumptions} were inserted if missing.
}
\description{
Add required data and assumptions to an environment
}
\details{
Note that data and assumptions are only assigned to the environment
if the corresponding attributes are not yet present. If already present,
they are not overwritten.
}
