% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelcomponent.R
\name{verify_is_component}
\alias{verify_is_component}
\title{Verify that an argument is a model_component for the right slot}
\usage{
verify_is_component(component, arg_name)
}
\arguments{
\item{component}{The object to check.}

\item{arg_name}{Name of the module function argument (equal to the slot).}
}
\description{
Used by the module functions to give a helpful error when
something else than a suitable component helper result is supplied.
}
\keyword{internal}
