% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{weighted.median}
\alias{weighted.median}
\title{Compute a weighted median of values}
\usage{
weighted.median(x, w)
}
\arguments{
\item{x}{A vector with values to be aggregated}

\item{w}{Weights for each entry}
}
\value{
Weighted median of the values in x
}
\description{
Compute a weighted median of values
}
