% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{withWarnings}
\alias{withWarnings}
\title{Collect warnings arising during computation of a result}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{Expression with the computation to run}
}
\value{
A \code{list}, with two elements: value and warnings
}
\description{
Collect warnings arising during computation of a result
}
\keyword{internal}
