% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{write_scalar_vars}
\alias{write_scalar_vars}
\title{Identify scalar variables in a stan model and write them to a text file}
\usage{
write_scalar_vars(modelfolder = "inst/stan", modelname = "EpiSewer_main.stan")
}
\arguments{
\item{modelfolder}{The folder containing the stan model file.}

\item{modelname}{The name of the stan model file.}
}
\description{
Identify scalar variables in a stan model and write them to a text file
}
\keyword{internal}
