% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{write_stan_digest}
\alias{write_stan_digest}
\title{Compute a digest of the stan model and write it to a text file}
\usage{
write_stan_digest(modelfolder = "inst/stan", modelname = "EpiSewer_main.stan")
}
\arguments{
\item{modelfolder}{The folder containing the stan model file.}

\item{modelname}{The name of the stan model file.}
}
\description{
Computes a digest (checksum) of the main stan model file and
all included stan files (e.g. functions), and writes it to
\code{stan_digest.txt} in the model folder. This digest is stored in every
\code{\link[=EpiSewerJob]{EpiSewerJob()}} created with the package, so that it can be verified
against the digest of the stan model compiled into a docker image (see
\code{\link[=sewer_pull_docker]{sewer_pull_docker()}}) before fitting a job in that image.
}
\keyword{internal}
