% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_assumptions_influenza_Zurich}
\alias{ww_assumptions_influenza_Zurich}
\title{Example sewer_assumptions object for Influenza A in Zurich}
\format{
A \code{sewer_assumptions} object as returned by \code{\link[=sewer_assumptions]{sewer_assumptions()}}.
}
\description{
Epidemiological assumptions for Influenza A, including
generation time and shedding load distributions derived from the
literature.
}
\seealso{
\code{\link{ww_data_influenza_Zurich}}
}
\keyword{datasets}
