% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_data_SARS_CoV_2_Zurich}
\alias{ww_data_SARS_CoV_2_Zurich}
\title{Example sewer_data object for SARS-CoV-2 in Zurich}
\format{
A \code{sewer_data} object as returned by \code{\link[=sewer_data]{sewer_data()}}.
}
\description{
Processed wastewater data for SARS-CoV-2 from the WWTP
Werdhölzli in Zürich (January–April 2022), subsampled to Monday and
Thursday measurements, ready for use with \code{\link[=EpiSewer]{EpiSewer()}}.
}
\seealso{
\code{\link{SARS_CoV_2_Zurich}}, \code{\link{ww_assumptions_SARS_CoV_2_Zurich}},
\code{\link{ww_result_SARS_CoV_2_Zurich}}
}
\keyword{datasets}
