% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{add_R_variability}
\alias{add_R_variability}
\title{Add change point model for the variability of Rt}
\usage{
add_R_variability(
  length_R,
  h,
  length_seeding,
  length_partial,
  changepoint_dist,
  modeldata
)
}
\arguments{
\item{length_R}{Length of the modeled Rt time series}

\item{changepoint_dist}{How many days should the change points be apart? If
zero, no change points are modeled (fixed R variability).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
Add change point model for the variability of Rt
}
\details{
The change points are placed going backwards from the end of the
time series, i.e. on the last day, then \code{changepoint_dist} days before
that, and so on. The distance between the first and second change point at
the start of the time series (partly falls into the seeding phase) varies
between half and double the \code{changepoint_dist}.
}
\keyword{internal}
