% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{component_functions}
\alias{component_functions}
\title{Get a list of modeling functions for a component}
\usage{
component_functions(component)
}
\arguments{
\item{component}{A character vector with the name of the component}
}
\value{
A character vector with all available modeling functions for the
component.
}
\description{
This function returns a list of all available modeling functions
for a given component.
}
