% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{concentrations_observe}
\alias{concentrations_observe}
\title{Observe concentration measurements}
\usage{
concentrations_observe(
  measurements = NULL,
  composite_window = 1,
  distribution = "normal",
  date_col = "date",
  concentration_col = "concentration",
  replicate_col = NULL,
  n_averaged = 1,
  n_averaged_col = NULL,
  total_droplets_col = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{measurements}{A \code{data.frame} with each row representing one
measurement. Must have at least a column with dates and a column with
concentration measurements.}

\item{composite_window}{Over how many days has each measured sample been
collected? If 1 (default), samples represent single days. If larger than 1,
samples are assumed to be equivolumetric composites over several dates. In
this case, the supplied dates represent the last day included in each
sample.}

\item{distribution}{Parametric distribution for concentration measurements.
Currently supported are "normal" (default and recommended), and
"lognormal". Distributions are truncated below zero if necessary.}

\item{date_col}{Name of the column containing the dates.}

\item{concentration_col}{Name of the column containing the measured
concentrations.}

\item{replicate_col}{Name of the column containing the replicate ID of each
measurement. This is used to identify multiple measurements made of a
sample from the same date. Should be \code{NULL} if only one measurement per
date was made.}

\item{n_averaged}{The number of replicates over which the measurements have
been averaged. This is typically used as an alternative to providing
several replicates per sample (i.e. the concentration provided in the
\code{measurements} \code{data.frame} is the arithmetic mean of several replicates).
Can be either a single number (it is then assumed that the number of
averaged replicates is the same for each observation) or a vector (one
value for each observation).}

\item{n_averaged_col}{Name of the column in the \code{measurements} data.frame
containing the number of replicates over which the measurements have been
averaged. This is an alternative to specifying \code{n_averaged}.}

\item{total_droplets_col}{Name of the column in the \code{measurements}
data.frame containing the total number of droplets in the ddPCR reaction
of each measurement. Only applies to concentration measurements obtain via
ddPCR. Can be used by the \code{\link[=noise_estimate_ddPCR]{noise_estimate_ddPCR()}} and
\code{\link[=LOD_estimate_ddPCR]{LOD_estimate_ddPCR()}} modeling components. Note that this is really the
\emph{total} number of droplets, not just the number of positive droplets.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option fits the \code{EpiSewer} model to pathogen concentrations
measured in wastewater samples. It is suitable for different quantification
methods such as qPCR or dPCR. The measured concentrations are modeled via a
lognormal likelihood.
}
\seealso{
Other {observation types}: 
\code{\link{droplets_observe}()}
}
\concept{{observation types}}
