% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{concentrations_observe}
\alias{concentrations_observe}
\title{Title}
\usage{
concentrations_observe(
  data,
  composite_window = 1,
  date_col = "date",
  concentration_col = "concentration",
  replicate_col = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicate_col}{}
}
\value{

}
\description{
Title
}
