% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{concentrations_observe_partitions}
\alias{concentrations_observe_partitions}
\title{Observe positive dPCR partition counts}
\usage{
concentrations_observe_partitions(
  measurements = NULL,
  composite_window = 1,
  date_col = "date",
  concentration_col,
  positive_partitions_col,
  replicate_col = NULL,
  n_averaged = 1,
  n_averaged_col = NULL,
  total_partitions_col,
  modeldata = modeldata_init()
)
}
\arguments{
\item{measurements}{A \code{data.frame} with each row representing one
measurement. Must have at least a column with dates and a column with
concentration measurements.}

\item{composite_window}{Over how many days has each measured sample been
collected? If 1 (default), samples represent single days. If larger than 1,
samples are assumed to be equivolumetric composites over several dates. In
this case, the supplied dates represent the last day included in each
sample.}

\item{date_col}{Name of the column containing the dates.}

\item{concentration_col}{Name of the column containing the measured
concentrations.}

\item{positive_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the number of positive partitions (e.g. positive
droplets for ddPCR) in the dPCR reaction of each measurement. If several
technical replicates are used, this should be the AVERAGE number of
positive partitions per replicate.}

\item{replicate_col}{Name of the column containing the replicate ID of each
measurement. This is used to identify multiple measurements made of a
sample from the same date. Should be \code{NULL} if only one measurement per
date was made.}

\item{n_averaged}{The number of technical replicates (i.e. repeated PCR runs)
used for each sample or biological replicate. The concentration provided in
the \code{measurements} \code{data.frame} is assumed to be the average or pooled
estimate from several technical replicates. Can be either a single number
(it is then assumed that the number of averaged replicates is the same for
each observation) or a vector (one value for each observation).}

\item{n_averaged_col}{Name of the column in the \code{measurements} data.frame
containing the number of technical replicates over which the measurements
have been averaged/pooled. This is an alternative to specifying
\code{n_averaged}.}

\item{total_partitions_col}{Name of the column in the \code{measurements}
data.frame containing the number of total partitions (e.g. droplets for
ddPCR) in the dPCR reaction of each measurement. If several technical
replicates are used, this should be the AVERAGE number of valid partitions
per replicate. Only applies when modeling concentration measurements via
the dPCR-specific noise model. Can be used by the \code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}}
and \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} modeling components. Note that this is really the
number of \emph{valid} partitions, not the number of positive partitions.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option fits the \code{EpiSewer} model to positive partition
counts in digital PCR (dPCR), e.g. positive droplets in ddPCR. This allows
the use of a dPCR-specific likelihood using a Binomial model for the number
of positive partitions observed. For a more generic likelihood, see
\code{\link[=concentrations_observe]{concentrations_observe()}}.
}
\seealso{
Other {observation types}: 
\code{\link{concentrations_observe}()}
}
\concept{{observation types}}
