% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{droplets_observe}
\alias{droplets_observe}
\title{Title}
\usage{
droplets_observe(
  data,
  composite_window = 1,
  date_col = "date",
  droplets_col = "droplets",
  replicate_col = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
