% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{droplets_observe}
\alias{droplets_observe}
\title{Observe positive ddPCR droplet counts}
\usage{
droplets_observe(
  data,
  composite_window = 1,
  date_col = "date",
  droplets_col = "droplets",
  replicate_col = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option fits the \code{EpiSewer} model to positive droplet counts
in digital droplet PCR (ddPCR) for the pathogen target of interest. This
allows the use of a ddPCR-specific likelihood using a Poisson distribution.
For a more generic likelihood, see \code{\link[=concentrations_observe]{concentrations_observe()}}.
}
\seealso{
Other {observation types}: 
\code{\link{concentrations_observe}()}
}
\concept{{observation types}}
