% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{EpiSewer}
\alias{EpiSewer}
\title{Estimate epidemiological parameters from wastewater measurements}
\usage{
EpiSewer(
  data = sewer_data(),
  assumptions = sewer_assumptions(),
  measurements = model_measurements(),
  sampling = model_sampling(),
  sewage = model_sewage(),
  shedding = model_shedding(),
  infections = model_infections(),
  forecast = model_forecast(),
  fit_opts = set_fit_opts(),
  results_opts = set_results_opts(),
  run_fit = TRUE
)
}
\arguments{
\item{data}{Observations such as concentration measurements and flows,
specified via \code{\link[=sewer_data]{sewer_data()}}. The data can also be directly supplied to the
relevant model components, in which case \code{data} can be left empty.}

\item{assumptions}{Assumptions about infections, shedding, or the sewage
system, specified via \code{\link[=sewer_assumptions]{sewer_assumptions()}}. The assumptions can also be
directly supplied to the relevant model components, in which case
\code{assumptions} can be left empty.}

\item{measurements}{The \code{measurements} module, see \code{\link[=model_measurements]{model_measurements()}}.}

\item{sampling}{The \code{sampling} module, see \code{\link[=model_sampling]{model_sampling()}}.}

\item{sewage}{The \code{sewage} module, see \code{\link[=model_sewage]{model_sewage()}}.}

\item{shedding}{The \code{shedding} module, see \code{\link[=model_shedding]{model_shedding()}}.}

\item{infections}{The \code{infections} module, see \code{\link[=model_infections]{model_infections()}}.}

\item{fit_opts}{Settings for model fitting, see \code{\link[=set_fit_opts]{set_fit_opts()}}.}

\item{results_opts}{Settings for results to be returned, see
\code{\link[=set_results_opts]{set_results_opts()}}.}

\item{run_fit}{If \code{TRUE} (the default), the model is fitted immediately. If
\code{FALSE}, the EpiSewerJob object is returned without fitting the model (it
can be fitted later or even on a different machine).}
}
\value{
If the model fitting was successful, a \code{list} with the following
elements is returned:
\itemize{
\item \code{job}: the \code{EpiSewerJob} that was run
\item \code{summary}: a summary of parameter estimates of interest
\item \code{fitted}: the fitted model (if \code{results_opts(fitted = TRUE)})
}

If the model fitting fails, a \code{list} with the following elements is
returned:
\itemize{
\item \code{error}: information about the errors and warnings that were thrown
\item \code{sampler_output}: potential outputs printed by the sampler
}
}
\description{
\code{EpiSewer} estimates the effective reproduction number Rt and
other parameters of interest from wastewater measurements over time. This
function combines data, assumptions and modeling details to fit a Bayesian
hierarchical model to the measurements. The resulting posterior estimates
of Rt and other parameters can be plotted and further analyzed.

The inputs to this function can be specified using various
helper functions (see below). It is best to define the inputs in advance as
separate variables and then pass them to \code{EpiSewer}.
}
