% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{EpiSewer}
\alias{EpiSewer}
\title{Title}
\usage{
EpiSewer(
  data = sewer_data(),
  assumptions = sewer_assumptions(),
  measurements = model_measurements(measurements = measurements_observe(measurements =
    data$measurements)),
  sampling = model_sampling(),
  sewage = model_sewage(flows = flows_observe(flows = data$flows)),
  shedding = model_shedding(incubation_dist = incubation_dist_assume(incubation_dist =
    assumptions$incubation_dist), shedding_dist = shedding_dist_assume(shedding_dist =
    assumptions$shedding_dist), load_per_case = load_per_case_assume(load_per_case =
    assumptions$load_per_case)),
  infections = model_infections(generation_dist =
    generation_dist_assume(generation_dist = assumptions$generation_dist)),
  model = get_stan_model(standata = standata),
  fit_opts = set_fit_opts(sampler = sampler_stan_mcmc(), fitted = TRUE)
)
}
\arguments{
\item{fit_opts}{}
}
\value{

}
\description{
Title
}
