% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{EpiSewerJob}
\alias{EpiSewerJob}
\title{Title}
\usage{
EpiSewerJob(
  job_name,
  model_def,
  modeldata,
  fit_opts,
  jobarray_size = 1,
  overwrite = TRUE,
  results_exclude = c()
)
}
\arguments{
\item{results_exclude}{}
}
\value{

}
\description{
Title
}
