% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{EpiSewerJob}
\alias{EpiSewerJob}
\title{Constructor for EpiSewerJob objects}
\usage{
EpiSewerJob(
  job_name,
  modeldata,
  fit_opts,
  results_opts,
  jobarray_size = 1,
  overwrite = TRUE,
  results_exclude = c()
)
}
\description{
Constructor for EpiSewerJob objects
}
