% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sewage.R
\name{flows_assume}
\alias{flows_assume}
\title{Assume a constant wastewater flow}
\usage{
flows_assume(flow_constant, modeldata = modeldata_init())
}
\arguments{
\item{flow_constant}{Daily wastewater flow volume, assumed to be the same for
all days.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option assumes a constant flow of wastewater at the
sampling site. Can be used as an approximation if no regular flow
measurements are available.
}
\details{
The flow volume unit should be the same as for the concentration
measurements, e.g. if concentrations are measured in gc/mL, then the flow
should be in mL as well.

Note that when the flow is unknown and therefore some arbitrary
value (e.g. \code{flow_constant=1}) is assumed, this must also be accounted for
in the assumed \code{load_per_case}. For this purpose, the function
\code{\link[=suggest_load_per_case]{suggest_load_per_case()}} offers a \code{flow_constant} argument.
}
\seealso{
Other {flow models}: 
\code{\link{flows_observe}()}
}
\concept{{flow models}}
