% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{flows_observe}
\alias{flows_observe}
\title{Title}
\usage{
flows_observe(
  standata = standata_init(),
  flows,
  date_col = "date",
  flow_col = "flow"
)
}
\arguments{
\item{flow_col}{}
}
\value{

}
\description{
Title
}
