% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sewage.R
\name{flows_observe}
\alias{flows_observe}
\title{Observe wastewater flows}
\usage{
flows_observe(
  data = NULL,
  date_col = "date",
  flow_col = "flow",
  modeldata = modeldata_init()
)
}
\arguments{
\item{data}{A \code{data.frame} with each row representing one day. Must have at
least a column with dates and a column with flow measurements.}

\item{date_col}{Name of the column containing the dates.}

\item{flow_col}{Name of the column containing the flows.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
This option accounts for daily wastewater flow volumes measured
at the sampling site. The flow can change due to rainfall or industrial
discharge, and directly influences pathogen concentrations in the
wastewater.
}
\details{
The flow volume unit should be the same as for the concentration
measurements, e.g. if concentrations are measured in gc/mL, then the flow
should be in mL as well.
}
\seealso{
Other {flow models}: 
\code{\link{flows_assume}()}
}
\concept{{flow models}}
