% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{generation_dist_assume}
\alias{generation_dist_assume}
\title{Title}
\usage{
generation_dist_assume(standata = standata_init(), generation_dist)
}
\arguments{
\item{generation_dist}{}
}
\value{

}
\description{
Title
}
