% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{generation_dist_assume}
\alias{generation_dist_assume}
\title{Assume a generation time distribution}
\usage{
generation_dist_assume(generation_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{generation_dist}{A numeric vector representing a discrete generation
time distribution, starting with the probability for a generation time of 1
day, 2 days, 3 days, and so on (a generation time of 0 days is excluded).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option assumes a fixed generation time distribution for the
renewal model in \code{EpiSewer}.
}
\details{
The generation time distribution here refers to the intrinsic
distribution of the time between infection of a primary case and infection
of its secondary cases. It is disease-specific and typically obtained from
literature.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_gamma_shifted]{get_discrete_gamma_shifted()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
