% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{get_checksums}
\alias{get_checksums}
\title{Get checksums that uniquely identify an EpiSewer job.}
\usage{
get_checksums(job, stanmodel_instance = NULL)
}
\arguments{
\item{job}{An EpiSewer job object.}

\item{stanmodel_instance}{A stanmodel instance.}
}
\value{
A \code{list} with checksums identifying the model, input, inits,
fit_opts and results_opts.
}
\description{
Get checksums that uniquely identify an EpiSewer job.
}
