% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dists.R
\name{get_discrete_lognormal}
\alias{get_discrete_lognormal}
\title{Get PMF of a discretized lognormal distribution.}
\usage{
get_discrete_lognormal(
  meanlog,
  sdlog,
  unit_mean = NULL,
  unit_sd = NULL,
  maxX,
  include_zero = TRUE,
  print_params = FALSE
)
}
\arguments{
\item{meanlog}{Mean of log}

\item{sdlog}{Standard deviation of log}

\item{unit_mean}{Alternative parameterization: unit scale mean}

\item{unit_sd}{Alternative parameterization: unit scale sd}

\item{maxX}{Right truncation point}

\item{include_zero}{Should the distribution explicitly cover X=0, or should
X=1 include the probability mass for X=0 too?}

\item{print_params}{Should the log-level parameters be printed?}
}
\value{
PMF of the discretized lognormal
}
\description{
This function accepts both log-scale and unit-scale parameters to specify the
lognormal distribution
}
