% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Title}
\usage{
get_stan_model(
  model_filename = NULL,
  modeldata = NULL,
  model_folder = "stan",
  profile = TRUE,
  threads = FALSE,
  force_recompile = FALSE
)
}
\arguments{
\item{force_recompile}{}
}
\value{

}
\description{
Title
}
