% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{horizon_assume}
\alias{horizon_assume}
\title{Specify the forecast horizon}
\usage{
horizon_assume(horizon, modeldata = modeldata_init())
}
\arguments{
\item{horizon}{The forecast horizon in days. If 0, no forecasts are produced.
Note that this functionality is intended for short-term forecasts.
Projections over longer horizons can be highly inaccurate.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option specifies a fixed forecast horizon in days. EpiSewer
will produce concentration predictions and forecasts of all latent
variables such as Rt, infections and load until the end of the forecast
horizon.
}
\details{
Forecasts account for the estimated variation of transmission
dynamics over time and therefore become more uncertain at longer forecast
horizons. However, it is important to keep in mind that in expectation the
model will project the current transmission dynamics to continue unchanged.
This assumption can be violated by various factors such as depletion of
susceptible individuals, changes in behavior, or public health
interventions.
}
