% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{horizon_none}
\alias{horizon_none}
\title{Do not produce forecasts}
\usage{
horizon_none(modeldata = modeldata_init())
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option specifies that no forecasts should be made. Only
estimates and concentration predictions until the last observed date are
produced.
}
