% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{incubation_dist_assume}
\alias{incubation_dist_assume}
\title{Title}
\usage{
incubation_dist_assume(incubation_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{incubation_dist}{}
}
\value{

}
\description{
Title
}
