% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{incubation_dist_assume}
\alias{incubation_dist_assume}
\title{Assume an incubation period distribution}
\usage{
incubation_dist_assume(incubation_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{incubation_dist}{A numeric vector representing a discrete incubation
period distribution, starting with the probability for an incubation period of 0
days, 1 day, 2 days, and so on.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
This option assumes a fixed incubation period distribution for
the shedding model in \code{EpiSewer}.
}
\details{
\code{EpiSewer} uses the incubation period as a proxy for the time
between infection and the start of shedding. This is because shedding load
distributions in the literature are often given from symptom onset onwards.
If the assumed shedding load distribution instead starts from the time of
infection, then use \code{incubation_dist=c(1)} (i.e. no lag).
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
