% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{infection_noise_estimate}
\alias{infection_noise_estimate}
\title{Title}
\usage{
infection_noise_estimate(
  standata = standata_init(),
  I_overdispersion = FALSE,
  I_xi_prior = stan_prior("I_xi", "normal", mu = 0, sigma = 1)
)
}
\arguments{
\item{I_xi_prior}{}
}
\value{

}
\description{
Title
}
