% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{infection_noise_estimate}
\alias{infection_noise_estimate}
\title{Title}
\usage{
infection_noise_estimate(
  overdispersion = FALSE,
  overdispersion_prior_mu = 0,
  overdispersion_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
