% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{load_per_case_assume}
\alias{load_per_case_assume}
\title{Title}
\usage{
load_per_case_assume(load_per_case = NULL, modeldata = modeldata_init())
}
\arguments{
\item{load_per_case}{}
}
\value{

}
\description{
Title
}
