% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{load_variation_estimate}
\alias{load_variation_estimate}
\title{Estimate individual-level variation in shedding loads}
\usage{
load_variation_estimate(
  cv_prior_mu = 0.1,
  cv_prior_sigma = 0.01,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}

\item{sd_prior_sigma}{}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Estimate individual-level variation in shedding loads
}
\seealso{
Other {load variation models}: 
\code{\link{load_variation_none}()}
}
\concept{{load variation models}}
