% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_assume}
\alias{LOD_assume}
\title{Assume a limit of detection}
\usage{
LOD_assume(
  limit = NULL,
  prob = 0.95,
  LOD_type = "exponential",
  modeldata = modeldata_init()
)
}
\arguments{
\item{limit}{Limit of detection. The concentration below which the pathogen
cannot be detected with sufficient probability, i.e. the measurement may be
zero although the pathogen is present in the sample.}

\item{prob}{What desired probability of detection does the limit refer to?
Default is 95\% (0.95): This means that the provided \code{limit} is the smallest
concentration at which the pathogen can still be detected with over 95\%
probability.}

\item{LOD_type}{The type of LOD model used. Currently, only "exponential" is
supported. This models an exponentially decreasing probability of zero
measurements / non-detection as a function of concentration. The
exponential model can be derived from the statistical properties of ddPCR,
but should also work well for other quantification methods such as qPCR.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
Pathogen concentrations below a certain threshold may not be
detectable and thus erroneously measured as 0. This option adjusts for a
known limit of detection and includes zero measurements in the likelihood.
}
\details{
The limit of detection is modeled using a hurdle model. In effect,
zero measurements provide a signal that the concentration in the respective
sample was likely below the limit of detection, but we don't know what the
exact concentration was.

The limit of detection is specific to the quantification approach
and protocol. It is usually established from a dedicated lab experiment.
}
\seealso{
{Visualize the assumed LOD as a function of concentration:}
\code{\link[=plot_LOD]{plot_LOD()}}

Other {LOD models}: 
\code{\link{LOD_estimate_ddPCR}()},
\code{\link{LOD_none}()}
}
\concept{{LOD models}}
