% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_assume}
\alias{LOD_assume}
\title{Assume a limit of detection}
\usage{
LOD_assume(limit = NULL, sharpness = 10, modeldata = modeldata_init())
}
\arguments{
\item{limit}{Limit of detection. The concentration below which measurements
will be determined as zero with substantial probability.}

\item{sharpness}{Sharpness of the threshold, see details.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Pathogen concentrations below a certain threshold may not be
detectable and thus erroneously measured as 0. This option adjusts for a
known limit of detection and includes zero measurements in the likelihood.
}
\details{
The limit of detection is modeled using a hurdle model. In effect,
zero measurements provide a signal that the concentration in the respective
sample was likely below the limit of detection, but we don't know what the
exact concentration was.

The limit of detection is highly specific to the quantification
approach and protocol. It is usually established from a dedicated lab
experiment.

\code{EpiSewer} does not model a clear-cut limit of detection but rather
a gradual increase in the probability of zero measurements as
concentrations become smaller. This is achieved using a sigmodial curve
that has its inflection point at the supplied \code{limit}. The \code{sharpness}
argument determines the steepness of this curve.
}
\seealso{
{Visualize the assumed limit of detection and its sharpness:}
\code{\link[=plot_LOD]{plot_LOD()}}

Other {LOD models}: 
\code{\link{LOD_none}()}
}
\concept{{LOD models}}
