% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{LOD_none}
\alias{LOD_none}
\title{Do not model a limit of detection}
\usage{
LOD_none(modeldata = modeldata_init())
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option drops all zero measurements from the likelihood and
does not explicitly model a limit of detection.
}
\details{
Dropping zero measurements implicitly assumes that any
level of concentration can theoretically lead to a zero measurement. Since
the corresponding observations are dropped, this will discard information,
but not bias estimates.
}
\seealso{
Other {LOD models}: 
\code{\link{LOD_assume}()},
\code{\link{LOD_estimate_ddPCR}()}
}
\concept{{LOD models}}
