% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{measurements_observe}
\alias{measurements_observe}
\title{Title}
\usage{
measurements_observe(
  standata = standata_init(),
  measurements,
  composite_window = 1,
  date_col = "date",
  measurement_col = "concentration"
)
}
\arguments{
\item{measurement_col}{}
}
\value{

}
\description{
Title
}
