% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{model_forecast}
\alias{model_forecast}
\title{Specify forecasting functionality}
\usage{
model_forecast(horizon = 0)
}
\arguments{
\item{horizon}{The forecasting horizon in days. If 0 (default), no forecasts
are produced.}
}
\value{
A \code{modeldata} object containing the data and specifications of the
\code{forecast} module.
}
\description{
Specify forecasting functionality
}
