% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{model_forecast}
\alias{model_forecast}
\title{Forecasting module}
\usage{
model_forecast(
  horizon = horizon_none(),
  damping = damping_assume(damping = 0.95)
)
}
\arguments{
\item{horizon}{The forecast horizon. How many days into the future should
EpiSewer forecast? Note that this functionality is intended for short-term
forecasts. Projections over longer horizons can be highly inaccurate.
Available options:
\itemize{
\item \code{\link[=horizon_none]{horizon_none()}}
\item \code{\link[=horizon_assume]{horizon_assume()}}
}}

\item{damping}{EpiSewer dampens the forecast of Rt so that trends in
transmission will level off after some time. This prevents unrealistic
extrapolation of transmission dynamics. Available options:
\itemize{
\item \code{\link[=damping_none]{damping_none()}}
\item \code{\link[=damping_assume]{damping_assume()}}
}}
}
\value{
A \code{modeldata} object containing the data and specifications of the
\code{forecast} module.
}
\description{
This module function is used to specify the components of the
\code{forecast} module in \code{EpiSewer}.

Each component can be specified using one or several helper
functions (see available options below). See the documentation of the
individual helper functions to adjust various settings.
}
\details{
Forecasts account for the estimated variation of transmission
dynamics over time and therefore tend to become more uncertain at longer
forecast horizons. However, it is important to keep in mind that depending
on the Rt model used, EpiSewer will project the current transmission
dynamics to continue unchanged (when using \code{\link[=R_estimate_rw]{R_estimate_rw()}},
\code{\link[=R_estimate_splines]{R_estimate_splines()}}, \code{\link[=R_estimate_piecewise]{R_estimate_piecewise()}}) or according to a
(dampened) linear trend (when using \code{\link[=R_estimate_ets]{R_estimate_ets()}} and
\code{\link[=R_estimate_changepoint_splines]{R_estimate_changepoint_splines()}}). This assumption can be violated by
various factors such as depletion of susceptible individuals, changes in
behavior, or public health interventions.
}
\seealso{
Other {module functions}: 
\code{\link{model_infections}()},
\code{\link{model_measurements}()},
\code{\link{model_sampling}()},
\code{\link{model_sewage}()},
\code{\link{model_shedding}()}
}
\concept{{module functions}}
