% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{model_forecast}
\alias{model_forecast}
\title{Forecasting module}
\usage{
model_forecast(horizon = horizon_none())
}
\arguments{
\item{horizon}{The forecast horizon. How many days into the future should
EpiSewer forecast? Note that this functionality is intended for short-term
forecasts. Projections over longer horizons can be highly inaccurate.
Available options:
\itemize{
\item \code{\link[=horizon_none]{horizon_none()}}
\item \code{\link[=horizon_assume]{horizon_assume()}}
}}
}
\value{
A \code{modeldata} object containing the data and specifications of the
\code{forecast} module.
}
\description{
This module function is used to specify the components of the
\code{forecast} module in \code{EpiSewer}.

Each component can be specified using one or several helper
functions (see available options below). See the documentation of the
individual helper functions to adjust various settings.
}
\details{
Forecasts account for the estimated variation of transmission
dynamics over time and therefore become more uncertain at longer forecast
horizons. However, it is important to keep in mind that in expectation the
model will project the current transmission dynamics to continue unchanged.
This assumption can be violated by various factors such as depletion of
susceptible individuals, changes in behavior, or public health
interventions.
}
