% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{model_infections}
\alias{model_infections}
\title{Title}
\usage{
model_infections(
  generation_dist,
  R = R_estimate_rw(),
  seeding = seeding_estimate(),
  infection_noise = infection_noise_estimate()
)
}
\arguments{
\item{infection_noise}{}
}
\value{

}
\description{
Title
}
