% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{model_measurements}
\alias{model_measurements}
\title{Title}
\usage{
model_measurements(
  concentrations = concentrations_observe(),
  noise = noise_estimate(),
  LOD = LOD_none()
)
}
\arguments{
\item{noise}{}
}
\value{

}
\description{
Title
}
