% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{model_shedding}
\alias{model_shedding}
\title{Title}
\usage{
model_shedding(
  incubation_dist = incubation_dist_assume(),
  shedding_dist = shedding_dist_assume(),
  load_per_case = load_per_case_assume()
)
}
\arguments{
\item{load_per_case}{}
}
\value{

}
\description{
Title
}
