% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{modeldata_init}
\alias{modeldata_init}
\title{Construct an unspecified EpiSewer model}
\usage{
modeldata_init()
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Construct an unspecified EpiSewer model
}
\examples{
modeldata_init()
}
