% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate}
\alias{noise_estimate}
\title{Estimate measurement noise}
\usage{
noise_estimate(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  type = "ddPCR",
  ddPCR_prior_droplets_mu = 20000,
  ddPCR_prior_droplets_sigma = 5000,
  ddPCR_droplets_fixed = TRUE,
  ddPCR_prior_scaling_mu = 1.5e-05,
  ddPCR_prior_scaling_sigma = 5e-06,
  ddPCR_scaling_fixed = FALSE,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements.}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations before the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations before the replication stage.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements. If multiple measurements (replicates) per sample are
provided, \code{EpiSewer} can also explicitly model variation before the
replication stage.
}
\details{
The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage:
\verb{Truncated normal}
}
}
