% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{noise_estimate}
\alias{noise_estimate}
\title{Title}
\usage{
noise_estimate(
  replicates = FALSE,
  sd_prior_mu = 0,
  sd_prior_sigma = 1,
  pre_replicate_sd_prior_mu = 0,
  pre_replicate_sd_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
