% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate}
\alias{noise_estimate}
\title{Estimate measurement noise}
\usage{
noise_estimate(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that when \code{replicates=TRUE}, this is only
the CV after the replication stage (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a constant coefficient of variation model.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.

For a non-constant coefficient of variation model, see
\code{\link[=noise_estimate_dPCR]{noise_estimate_dPCR()}}.
}
\details{
When \code{replicates=TRUE}, two coefficients of variation are estimated:
\itemize{
\item the CV before the replication stage (see \code{pre_replicate_cv_prior_mu})
\item the CV after the replication stage (see \code{cv_prior_mu})
}

The meaning of these CV estimates depends on the type of replicates. If the
replicates are biological replicates (i.e. independently processed), then
\code{cv} estimates the noise in the preprocessing and PCR, and
\code{pre_replicate_cv} estimates the noise from anything before preprocessing
(e.g. sampling noise and all other unexplained variation). In contrast, if
the replicates are technical replicates (i.e. several PCR runs of the same
preprocessed sample), then \code{cv} estimates only the PCR noise, and
\code{pre_replicate_cv} estimates all other noise (including preprocessing
noise.)

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage (\code{pre_replicate_cv}):
\verb{Truncated normal}
}
}
\seealso{
Other \link{noise models}: 
\code{\link{noise_estimate_constant_var}()},
\code{\link{noise_estimate_dPCR}()}
}
\concept{\link{noise models}}
