% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_constant_var}
\alias{noise_estimate_constant_var}
\title{Estimate measurement noise with constant variance}
\usage{
noise_estimate_constant_var(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  warn = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that when \code{replicates=TRUE}, this is only
the CV after the replication stage (see details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a constant variance model. This is usually a
misspecification and is only supported for comparison purposes.

For a constant coefficient of variation model, see
\code{\link[=noise_estimate]{noise_estimate()}}, and for a non-constant coefficient of variation model,
see \code{\link[=noise_estimate_ddPCR]{noise_estimate_ddPCR()}}.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
Note that although this model keeps the variance constant, the prior
for the measurement noise is still in terms of the (average) coefficient of
variation (CV). This makes prior specification easier since it the CV is
unitless.

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage:
\verb{Truncated normal}
}
}
\seealso{
Other \link{noise models}: 
\code{\link{noise_estimate_ddPCR}()},
\code{\link{noise_estimate}()}
}
\concept{\link{noise models}}
