% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_ddPCR}
\alias{noise_estimate_ddPCR}
\title{Estimate measurement noise for digital droplet PCR data}
\usage{
noise_estimate_ddPCR(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  ddPCR_prior_droplets_mu = 20000,
  ddPCR_prior_droplets_sigma = 5000,
  ddPCR_droplets_fixed = TRUE,
  ddPCR_prior_scaling_mu = 1.5e-05,
  ddPCR_prior_scaling_sigma = 5e-06,
  ddPCR_scaling_fixed = FALSE,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements.}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{ddPCR_prior_droplets_mu}{Prior (mean) on the number of droplets in the
ddPCR reaction.}

\item{ddPCR_prior_droplets_sigma}{Prior (standard deviation) on the number of
droplets in the ddPCR reaction.}

\item{ddPCR_droplets_fixed}{If TRUE (default), the number of droplets is
fixed to the prior mean and not estimated. This is recommended if no
replicates are available.}

\item{ddPCR_prior_scaling_mu}{Prior (mean) on the concentration scaling
factor for the ddPCR reaction. The concentration scaling factor is the
droplet volume, scaled by the dilution of the wastewater in the ddPCR
reaction. See details for further explanation.}

\item{ddPCR_prior_scaling_sigma}{Prior (standard deviation) on the
concentration scaling factor for the ddPCR reaction.}

\item{ddPCR_scaling_fixed}{If TRUE, the concentration scaling factor is fixed
to the prior mean and not estimated.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations before the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations before the replication stage.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a coefficient of variation model specialised for digital
droplet PCR. Specifically, the coefficient of variation is modeled as a
function of the expected concentration according to the statistical
properties of ddPCR.

This model predicts a higher coefficient of variation at smaller
concentrations, which often leads to a better model fit, even for
measurements from other quantification methods such as qPCR.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
The concentration scaling factor (see \code{ddPCR_prior_scaling_mu},
\code{ddPCR_prior_scaling_sigma}, \code{ddPCR_scaling_fixed}) is the droplet volume,
scaled by the dilution of the wastewater in the ddPCR reaction. The
dilution accounts for all extraction and preparation steps. For example, if
the droplet volume is 4.5e-7 mL and the dilution of the wastewater is 100:3
(i.e. 100 gc/mL in the original wastewater sample correspond to 3 gc/mL in
the PCR reaction), then the overall scaling factor is 4.5e-7 * 100 / 3 =
1.5e-5.

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements: \verb{Truncated normal}
\item prior for number of droplets in ddPCR: \verb{Truncated normal}
\item prior for concentration scaling factor of ddPCR: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage:
\verb{Truncated normal}
}
}
\seealso{
\link{LOD_estimate_ddPCR} for a limit of detection model specialised for
ddPCR.

Other \link{noise models}: 
\code{\link{noise_estimate}()}
}
\concept{\link{noise models}}
