% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_measurements.R
\name{noise_estimate_dPCR}
\alias{noise_estimate_dPCR}
\title{Estimate measurement noise for digital PCR data}
\usage{
noise_estimate_dPCR(
  replicates = FALSE,
  cv_prior_mu = 0,
  cv_prior_sigma = 1,
  total_partitions_prior_mu = 20000,
  total_partitions_prior_sigma = 5000,
  total_partitions_observe = FALSE,
  partition_variation_prior_mu = 0,
  partition_variation_prior_sigma = 0.05,
  volume_scaled_prior_mu = 1.5e-05,
  volume_scaled_prior_sigma = 5e-06,
  pre_replicate_cv_prior_mu = 0,
  pre_replicate_cv_prior_sigma = 1,
  prePCR_noise_type = "log-normal",
  use_taylor_approx = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{replicates}{Should replicates be used to explicitly model variation
before the replication stage?}

\item{cv_prior_mu}{Prior (mean) on the coefficient of variation of
concentration measurements. Note that in contrast to using
\code{\link[=noise_estimate]{noise_estimate()}}, this does \emph{not} include the technical noise of the
digital PCR. This is because the dPCR noise is explicitly modeled (using
the number of partitions and conversion factor). Moreover, when
\code{replicates=TRUE}, this is only the CV after the replication stage (see
details for more explanation).}

\item{cv_prior_sigma}{Prior (standard deviation) on the coefficient of
variation of concentration measurements.}

\item{total_partitions_prior_mu}{Prior (mean) on the total number of
partitions in the dPCR reaction.}

\item{total_partitions_prior_sigma}{Prior (standard deviation) on the total
number of partitions in the dPCR reaction. If this is set to zero, the
total number of partitions will be fixed to the prior mean and not
estimated.}

\item{total_partitions_observe}{If TRUE, the total number of partitions is
taken from the supplied measurements \code{data.frame}. This requires that the
argument \code{total_partitions_col} is specified in \code{\link[=concentrations_observe]{concentrations_observe()}}.}

\item{partition_variation_prior_mu}{Prior (mean) on the coefficient of
variation of the total number of partitions in the dPCR reaction. Usually,
the maximum number of partitions possible for a given dPCR chip is not
reached, i.e. a certain number of partitions is lost. This loss varies
between PCR runs, and is modeled as log-normal distributed in EpiSewer.}

\item{partition_variation_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of the total number of partitions in the dPCR
reaction.}

\item{volume_scaled_prior_mu}{Prior (mean) on the conversion factor
(partition volume scaled by the dilution of wastewater in the assay) for
the dPCR reaction. See details for further explanation.}

\item{volume_scaled_prior_sigma}{Prior (standard deviation) on the conversion
factor (partition volume scaled by the dilution of wastewater in the assay)
for the dPCR reaction. If this is set to zero, the conversion factor will
be fixed to the prior mean and not estimated.}

\item{pre_replicate_cv_prior_mu}{Prior (mean) on the coefficient of variation
of concentrations \emph{before} the replication stage.}

\item{pre_replicate_cv_prior_sigma}{Prior (standard deviation) on the
coefficient of variation of concentrations \emph{before} the replication stage.}

\item{prePCR_noise_type}{The parametric distribution to assume for noise
before the PCR assay. Currently supported are "log-normal" and "gamma". The
choice of the parametric distribution typically makes no relevant
difference for the noise model, but can make a relevant difference for the
LOD model if \code{\link[=LOD_estimate_dPCR]{LOD_estimate_dPCR()}} is used.}

\item{use_taylor_approx}{If TRUE (default), a Taylor expansion approximation
is used to estimate the CV of measurements under pre-PCR noise. The
approximation is very accurate, unless concentrations are extremely high
(so high that the quality of the measurements from dPCR would anyway be
questionable).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the unexplained variation in wastewater
measurements using a coefficient of variation model specialized for digital
PCR (e.g. digital droplet PCR). Specifically, the coefficient of variation
is modeled as a function of the expected concentration according to the
statistical properties of dPCR.

This model predicts a higher coefficient of variation at smaller
concentrations, which often leads to a better model fit, even for
measurements from other quantification methods such as qPCR.

If multiple measurements (replicates) per sample are provided,
\code{EpiSewer} can also explicitly model variation before the replication
stage.
}
\details{
The conversion factor (see \code{volume_scaled_prior_mu},
\code{volume_scaled_prior_sigma}) is the partition volume scaled by the dilution
of the wastewater in the assay. The dilution accounts for all
extraction and preparation steps. For example, if the partition volume is
4.5e-7 mL and the dilution of the wastewater is 100:3 (i.e. 100 gc/mL in
the original wastewater sample correspond to 3 gc/mL in the PCR reaction),
then the overall conversion factor is 4.5e-7 * 100 / 3 = 1.5e-5.

When \code{replicates=TRUE}, two coefficients of variation are estimated:
\itemize{
\item the CV before the replication stage (see \code{pre_replicate_cv_prior_mu})
\item the CV after the replication stage (see \code{cv_prior_mu})

The meaning of these CV estimates depends on the type of replicates. If the
replicates are biological replicates (i.e. independently processed), then
\code{cv} estimates the noise in the preprocessing before the PCR, and
\code{pre_replicate_cv} estimates the noise from anything before preprocessing
(e.g. sampling noise and all other unexplained variation). In contrast, if
the replicates are technical replicates (i.e. several PCR runs of the same
preprocessed sample), then \code{cv} estimates only unexplained PCR noise
(should be close to zero), and \code{pre_replicate_cv} estimates all other noise
(including preprocessing noise.)
}

The priors of this component have the following functional form:
\itemize{
\item coefficient of variation of concentration measurements (\code{cv}): \verb{Truncated normal}
\item mean number of partitions in dPCR: \verb{Truncated normal}
\item coefficient of variation of number of partitions in dPCR: \verb{Truncated normal}
\item conversion factor for dPCR: \verb{Truncated normal}
\item coefficient of variation of concentration before the replication stage (\code{pre_replicate_cv}):
\verb{Truncated normal}
}
}
\seealso{
\link{LOD_estimate_dPCR} for a limit of detection model specialised for
dPCR.

Other \link{noise models}: 
\code{\link{noise_estimate_constant_var}()},
\code{\link{noise_estimate}()}
}
\concept{\link{noise models}}
