% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{outliers_none}
\alias{outliers_none}
\title{Do not model outliers}
\usage{
outliers_none(modeldata = modeldata_init())
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option does not model outliers in sampled concentrations.
}
\seealso{
Other {outlier models}: 
\code{\link{outliers_estimate}()}
}
\concept{{outlier models}}
