% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Plot predicted concentration}
\usage{
plot_concentration(
  results = NULL,
  measurements = NULL,
  include_noise = TRUE,
  median = FALSE,
  mark_outliers = FALSE,
  forecast = TRUE,
  forecast_horizon = NULL,
  date_margin_left = 0,
  date_margin_right = 1,
  facet_models = FALSE,
  base_model = "",
  model_levels = NULL,
  concentration_col = "concentration",
  date_col = "date",
  outlier_col = "is_outlier",
  type = "time"
)
}
\arguments{
\item{results}{Results object returned by \code{\link[=EpiSewer]{EpiSewer()}} after model fitting.
Can also be a named \code{list} with results from different model runs, in which
case all results are plotted together and distinguished by colors.}

\item{measurements}{A \code{data.frame} with observed measurements, which will be
plotted alongside the predicted values (useful to assess model fit). Can
also be used with \code{results = NULL}, in which case only the observed
measurements are plotted.}

\item{include_noise}{If \code{TRUE} (default), concentrations including
measurement noise are shown. If \code{FALSE}, only the expected concentrations
are shown.}

\item{median}{Should the estimated median be shown, or only the credible
intervals? Default is \code{FALSE} to avoid over-interpretation of the median.}

\item{mark_outliers}{If \code{TRUE}, outliers in the \code{measurements} are
highlighted in red. See also argument \code{outlier_col} below.}

\item{forecast}{Should forecasted concentrations be shown? Default is true.
This requires that the model was fitted with a forecast horizon, see
\code{\link[=model_forecast]{model_forecast()}}. Because concentrations depend on the flow volume,
accurate concentration forecasts are only possible if flow values beyond
the estimation date are provided. This is typically only possible when
estimating retrospectively. If flow values beyond the estimation date are
not available, the forecasted concentrations will be based on the median
flow volume.}

\item{forecast_horizon}{How many days into the future should forecasts be
plotted? Note that this is restricted by the forecast horizon specified
during model fitting, see \code{\link[=horizon_assume]{horizon_assume()}}.}

\item{date_margin_left}{By how many days into the past should the plot be
expanded? Can also be negative to cut off some of the earliest dates.}

\item{date_margin_right}{By how many days into the future should the plot be
expanded? Can also be negative to cut off some of the latest dates. By
default, this is 1 to improve the visibility of the latest date.}

\item{facet_models}{Should the plot be faceted by model? Default is \code{FALSE}.}

\item{base_model}{Name of the base model (in the named list provided to
\code{results}) which should be compared to the other models. This model will be
plotted in black and will not be part of the legend.}

\item{model_levels}{A \code{character} vector with the names of the models to be
included. The colors and legend will be ordered according to the order in
\code{model_levels}.}

\item{concentration_col}{Name of the column in the measurements \code{data.frame}
which contains the measured concentration that should be plotted.}

\item{date_col}{Name of the date column in the measurements \code{data.frame}.}

\item{outlier_col}{Name of a logical column in the measurements \code{data.frame}
which identifies outlier measurements (for example added by
\code{\link[=mark_outlier_spikes_median]{mark_outlier_spikes_median()}}).}

\item{type}{If "time" (default), the concentration time series is plotted. If
"pp_check", the observations are ordered by concentration and plotted
against the predicted concentration (useful for posterior predictive
checks).}
}
\value{
A ggplot object showing predicted and observed concentrations over
time. Can be further manipulated using \link{ggplot2} functions to adjust themes
and scales, and to add further geoms.
}
\description{
Plots measured and/or predicted concentrations over time. The
predicted concentrations are taken from a fitted \code{EpiSewer} model.
}
\details{
When plotting a posterior predictive check (\code{type="pp_check"}), each
observed concentration is shown together with the 95\% Credible Interval of
the posterior predictive distribution for that observation. The
observations are ordered by concentration, which helps to visualize bias
and variance of the predictions as a function of the concentration.
}
