% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_load}
\alias{plot_load}
\title{Plot the estimated load}
\usage{
plot_load(results, median = FALSE)
}
\arguments{
\item{results}{Results object returned by \code{\link[=EpiSewer]{EpiSewer()}} after model fitting.}

\item{median}{Should the estimated median be shown, or only the credible
intervals? Default is \code{FALSE} to avoid over-interpretation of the median.}
}
\value{
A ggplot object showing the estimated load over time. Can be further
manipulated using \link{ggplot2} functions to adjust themes and scales, and add
further geoms.
}
\description{
Plots the estimated load in wastewater over time from a fitted
\code{EpiSewer} model.
}
