% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_plot.R
\name{plot_sample_effects}
\alias{plot_sample_effects}
\title{Plot estimated sample effects}
\usage{
plot_sample_effects(results)
}
\arguments{
\item{results}{Results object returned by \code{\link[=EpiSewer]{EpiSewer()}} after model fitting.}
}
\value{
A ggplot object showing the estimated effect sizes. Can be further
manipulated using \link{ggplot2} functions to adjust themes and scales, and add
further geoms.
}
\description{
Plots estimated effect sizes for sample covariates with 95\%
credible intervals. Only works if the \code{EpiSewer} model
included sample covariates, see e.g. \code{\link[=sample_effects_estimate_matrix]{sample_effects_estimate_matrix()}}.
}
