% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_ets}
\alias{R_estimate_ets}
\title{Smooth Rt estimates via exponential smoothing}
\usage{
R_estimate_ets(
  level_prior_mu = 1,
  level_prior_sigma = 0.8,
  trend_prior_mu = 0,
  trend_prior_sigma = 0.1,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.05,
  smooth_fixed = NULL,
  smooth_prior_shapes = c(50, 50),
  trend_smooth_fixed = NULL,
  trend_smooth_prior_shapes = c(50, 50),
  dampen_fixed = 0.9,
  dampen_prior_shapes = c(50, 5),
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{noncentered}{}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Smooth Rt estimates via exponential smoothing
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
