% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_ets}
\alias{R_estimate_ets}
\title{Title}
\usage{
R_estimate_ets(
  level_prior_mu = 1,
  level_prior_sigma = 0.8,
  trend_prior_mu = 0,
  trend_prior_sigma = 0.1,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.05,
  smooth_fixed = NULL,
  smooth_prior_shapes = c(50, 50),
  trend_smooth_fixed = NULL,
  trend_smooth_prior_shapes = c(50, 50),
  dampen_fixed = 0.9,
  dampen_prior_shapes = c(50, 5),
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
