% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_ets}
\alias{R_estimate_ets}
\title{Title}
\usage{
R_estimate_ets(
  standata = standata_init(),
  R_level_start_prior = stan_prior("R_level_start", "normal", mu = 1, sigma = 0.8),
  R_trend_start_prior = stan_prior("R_trend_start", "normal", mu = 0, sigma = 0.1),
  R_sd_prior = stan_prior("R_sd", "half-normal", mu = 0, sigma = 0.05),
  ets_diff = FALSE,
  ets_noncentered = TRUE,
  ets_alpha_fixed = NULL,
  ets_alpha_prior = c(50, 50),
  ets_beta_fixed = NULL,
  ets_beta_prior = c(50, 50),
  ets_phi_fixed = 0.9,
  ets_phi_prior = c(50, 5)
)
}
\arguments{
\item{ets_phi_prior}{}
}
\value{

}
\description{
Title
}
