% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_ets}
\alias{R_estimate_ets}
\title{Estimate Rt via exponential smoothing}
\usage{
R_estimate_ets(
  level_prior_mu = 1,
  level_prior_sigma = 0.8,
  trend_prior_mu = 0,
  trend_prior_sigma = 0.1,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.1,
  sd_changepoint_dist = 28,
  sd_changepoint_sd = 0.025,
  link = "inv_softplus",
  R_max = 6,
  smooth_prior_mu = 0.5,
  smooth_prior_sigma = 0.05,
  trend_smooth_prior_mu = 0.5,
  trend_smooth_prior_sigma = 0.05,
  dampen_prior_mu = 0.9,
  dampen_prior_sigma = 0,
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{level_prior_mu}{Prior (mean) on the initial level of Rt.}

\item{level_prior_sigma}{Prior (standard deviation) on the initial level of
Rt.}

\item{trend_prior_mu}{Prior (mean) on the initial trend of Rt.}

\item{trend_prior_sigma}{Prior (standard deviation) on the initial trend of
Rt.}

\item{sd_prior_mu}{Prior (mean) on the standard deviation of the innovations.}

\item{sd_prior_sigma}{Prior (standard deviation) on the standard deviation of
the innovations.}

\item{sd_changepoint_dist}{The variability of Rt can change over time, e.g.
during the height of an epidemic wave, countermeasures may lead to much
faster changes in Rt than observable at other times. This potential
variability is accounted for using change points placed at regular
intervals. The standard deviation of the state space model innovations then
evolves linearly between the change points. The default change point
distance is 28 days. If set to zero, no change points are modeled.}

\item{sd_changepoint_sd}{This parameter controls the variability of the change
points. When change points are modeled, EpiSewer will estimate a baseline
standard deviation (see \code{sd_prior_mu} and \code{sd_prior_sigma}), and model
change point values as independently distributed with mean equal to this
baseline and standard deviation \code{sd_changepoint_sd}.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{smooth_prior_mu}{Prior (mean) on the smoothing parameter. Must be
between 0 and 1.}

\item{smooth_prior_sigma}{Prior (standard deviation) on the smoothing
parameter. If this is set to zero, the smoothing parameter will be fixed to
\code{smooth_prior_mu} and not estimated. If positive, a beta prior with the
corresponding mean and standard deviation is used.}

\item{trend_smooth_prior_mu}{Prior (mean) on the trend smoothing parameter.
Must be between 0 and 1.}

\item{trend_smooth_prior_sigma}{Prior (standard deviation) on the trend
smoothing parameter. If this is set to zero, the trend smoothing parameter
will be fixed to \code{trend_smooth_prior_mu} and not estimated. If positive, a
beta prior with the corresponding mean and standard deviation is used.}

\item{dampen_prior_mu}{Prior (mean) on the dampening parameter. Must be
between 0 and 1.}

\item{dampen_prior_sigma}{Prior (standard deviation) on the dampening
parameter. If this is set to zero, the dampening parameter will be fixed to
\code{dampen_prior_mu} and not estimated. If positive, a beta prior with the
corresponding mean and standard deviation is used.}

\item{differenced}{If \code{FALSE} (default), exponential smoothing is applied to
the absolute Rt time series. If \code{TRUE}, it is instead applied to the
differenced time series. This makes the level become the trend, and the
trend become the curvature.}

\item{noncentered}{If \code{TRUE} (default), a non-centered parameterization is
used to model the innovations in the state space process (for better
sampling efficiency).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number over time
using exponential smoothing. It implements Holt's linear trend method with
dampening through an innovations state space model with a level, trend, and
dampening component.
}
\details{
The innovations state space model consists of three components: a
level, a trend, and a dampening component.
\itemize{
\item The level is smoothed based on the levels from earlier time steps,
with exponentially decaying weights, as controlled by a smoothing parameter
(often called alpha). Note that \emph{smaller} values of \code{alpha} indicate
\emph{stronger} smoothing. In particular, \code{alpha = 1} means that only the last
level is used.
\item The trend is smoothed based on the trends from earlier time steps,
with exponentially decaying weights, as controlled by a trend smoothing
parameter (often called beta). Note that \emph{smaller} values of \code{beta} indicate
\emph{stronger} smoothing. In particular, \code{beta = 1} means that only the last
trend is used.
\item The dampening determines how long a previous trend continues into the
future before it levels of to a stationary time series. The strength of
dampening is controlled by a dampening parameter (often called phi). Note
that \emph{smaller} values of \code{phi} indicate \emph{stronger} dampening. In particular,
\code{phi = 1} means no dampening. Values below \code{phi = 0.8} are seldom in
practice as the dampening becomes very strong.
}

Often, \code{alpha}, \code{beta}, and \code{phi} are jointly unidentifiable. It may
therefore be necessary to fix at least one of the parameters (typically
\code{phi}) or supply strong priors.

Note that the smoothness of retrospective Rt estimates is often more
influenced by the prior on the standard deviation of innovations than the
smoothing and trend smoothing parameters. The smoothing parameters mostly
have an influence on estimates close to the present / date of estimation,
when limited data signal is available. Here, the standard deviation of the
innovations influences how uncertain Rt estimates are close to the present.

The priors of this component have the following functional form:
\itemize{
\item initial level of Rt: \code{Normal}
\item initial trend of Rt: \code{Normal}
\item standard deviation of innovations: \verb{Truncated normal}
\item smoothing parameter: \code{Beta}
\item trend smoothing parameter: \code{Beta}
\item dampening parameter: \code{Beta}
}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
