% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_gp}
\alias{R_estimate_gp}
\title{Estimate Rt using Gaussian processes}
\usage{
R_estimate_gp(
  R_intercept_prior_mu = 1,
  R_intercept_prior_sigma = 0,
  length_scale_prior_mu = 7 * 3,
  length_scale_prior_sigma = 7/2,
  magnitude_prior_mu = 0.2,
  magnitude_prior_sigma = 0.05,
  long_length_scale_prior_mu = 7 * 4 * 3,
  long_length_scale_prior_sigma = 7,
  long_magnitude_prior_mu = 0.4,
  long_magnitude_prior_sigma = 0.1,
  matern_nu = c(3/2, 5/2, 1/2),
  boundary_factor = 3,
  n_basis_factor = 3.42,
  link = "inv_softplus",
  R_max = 6,
  modeldata = modeldata_init()
)
}
\arguments{
\item{R_intercept_prior_mu}{Prior (mean) for the intercept of Rt. Should be
set to 1 unless you have a clear a priori expectation of the average Rt
during the modeled time period.}

\item{R_intercept_prior_sigma}{Prior (standard deviation) for the intercept
of Rt. By default, we fix \code{R_intercept} to 1 by setting
\code{R_intercept_prior_sigma=0}. This is because deviations from Rt=1 are
already captured by the long-term trend component (see below).}

\item{length_scale_prior_mu}{Prior (mean) on the length scale of the
short-term Gaussian process (in days). This influences the smoothness of
Rt. A higher length scale means that the Rt will change more slowly.
Choosing a length scale that is too short can lead to overfitting of the Rt
trajectory, while choosing a length scale that is too long can result in
unrealistically smooth Rt estimates.}

\item{length_scale_prior_sigma}{Prior (standard deviation) on the length
scale of the short-term Gaussian process. Set to zero to fix the length
scale.}

\item{magnitude_prior_mu}{Prior (mean) on the magnitude of the short-term
Gaussian process. Can be approximately interpreted as the marginal standard
deviation of Rt. A higher magnitude allows more extreme Rt values.}

\item{magnitude_prior_sigma}{Prior (standard deviation) on the magnitude of
the short-term Gaussian process. Set to zero to fix the magnitude.}

\item{long_length_scale_prior_mu}{Prior (mean) on the length scale of the
long-term Gaussian process (in days). This should be quite long, at least
several times the mean shedding delay of the pathogen, and significantly
larger than the mean prior for the short-term GP (\code{length_scale_prior_mu}).}

\item{long_length_scale_prior_sigma}{Prior (standard deviation) on the length
scale of the long-term Gaussian process. Set to zero to fix the length
scale.}

\item{long_magnitude_prior_mu}{Prior (mean) on the magnitude of the long-term
Gaussian process. Can be approximately interpreted as the marginal standard
deviation of the long-term Rt trend. A higher magnitude allows more extreme
Rt values.}

\item{long_magnitude_prior_sigma}{Prior (standard deviation) on the magnitude
of the long-term Gaussian process. Set to zero to fix the magnitude.}

\item{matern_nu}{The smoothness parameter of the Matern kernel. The default
is 3/2, other possible choices are 5/2 (more smooth) and 1/2 (less smooth).
However, we recommend tuning smoothness primarily using the length scale
priors.}

\item{boundary_factor}{The boundary factor used in the Gaussian process
approximation. The default (\code{boundary_factor = 3}) is higher than the
minimum recommendation from Riutort-Mayol et al. to ensure accurate
real-time estimation. Lower values can lead to boundary effects close to
the present and start of the time series and are thus not recommended. When
modeling very long shedding delays, the boundary factor might need to be
further increased. Note that this will also automatically increase the
number of basis functions used and thereby slow down sampling.}

\item{n_basis_factor}{Factor used to automatically determine \code{m}, the number
of basis functions in the Gaussian process approximation. Based on
recommendations from Riutort-Mayol et al., we let \code{m = n_basis_factor * c / (l / S)}, where \code{c} is the \code{boundary_factor}, \code{l} is the length scale of
the GP (we use the 5\% quantile of \code{gp_length_prior} for a conservative
result), and \code{S} is the maximum absolute value of the zero-centered input
space, which is \code{(n-1)/2} where \code{n} is the number of Rt time steps modeled.
This means that the number of basis function automatically adapts to the
length scale and number of observations in the model. For the default
settings (\code{boundary_factor = 3} and \code{n_basis_factor = 3.42}) and a lower
length scale of 3 weeks, this corresponds to approx. 0.25x the number of Rt
time points modeled. Increasing \code{n_basis_factor} will make the
approximation more accurate by proportionally increasing the number of
basis functions, but can slow down sampling.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number Rt over
time using a Gaussian process (GP) model. There are two GPs: one for the
long-term trend in Rt, and one for short-term deviations from this trend.
}
\details{
The estimated Rt trajectory is primarily influenced by the priors
for the \emph{length scale} and the \emph{magnitude} of the short-term and long-term
Gaussian processes. We recommend adjusting the \emph{magnitude} when maximum Rt
values seem to be too low or too high. In contrast, we recommend adjusting
the \emph{length scale} when the Rt trajectory seems to have too little
resolution (try shorter length scales) or is overfitting on noise (try
longer length scales). Note that, to ensure identifiability, the prior mean
for the length scale of the long-term GP should always be significantly
larger than the prior mean for the short-term GP.

The Gaussian process is modeled using a Hilbert space approximation
as described in Riutort-Mayol et al. (2023). This allows for fast inference
without significant loss of accuracy. See the reference for more detail on
choosing an adequate boundary factor and basis function factor. The
\code{EpiSewer} implementation of the approximate Gaussian process is strongly
inspired by the implementation in the
\href{https://github.com/epiforecasts/EpiNow2/}{EpiNow2} package.

The priors of this component have the following functional form:
\itemize{
\item R_intercept (intercept):
\code{Normal}
\item magnitude (magnitude):
\verb{Truncated Normal}
\item length_scale (length scale):
\verb{Truncated Normal}
\item long_magnitude (magnitude for long-term trend):
\verb{Truncated Normal}
\item long_length_scale (length scale for long-term trend):
\verb{Truncated Normal}
}
}
\references{
Riutort-Mayol, G., Bürkner, PC., Andersen, M.R. et al. Practical
Hilbert space approximate Bayesian Gaussian processes for probabilistic
programming. \emph{Stat Comput} 33, 17 (2023).
\url{https://doi.org/10.1007/s11222-022-10167-2}

Abbott S, Hellewell J, Thompson RN et al. Estimating the
time-varying reproduction number of SARS-CoV-2 using national and
subnational case counts. \emph{Wellcome Open Res} 5, 112 (2020).
\url{https://doi.org/10.12688/wellcomeopenres.16006.2}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_changepoint_splines}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_piecewise}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_smooth_derivative}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
