% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_gp}
\alias{R_estimate_gp}
\title{Estimate Rt using a Gaussian process}
\usage{
R_estimate_gp(
  R_start_prior_mu = 1,
  R_start_prior_sigma = 0.8,
  gp_sigma_prior_mu = 0.05,
  gp_sigma_prior_sigma = 0.01,
  gp_length_prior_mu = 28,
  gp_length_prior_sigma = 3.5,
  gp_length_max = 7 * 4 * 2,
  gp_matern_nu = c(3/2, 5/2, 1/2),
  gp_boundary_factor = 3,
  gp_n_basis_factor = 3.42,
  phi = 0.95,
  link = "inv_softplus",
  R_max = 6,
  modeldata = modeldata_init()
)
}
\arguments{
\item{R_start_prior_mu}{Prior (mean) on the initial reproduction number
(intercept).}

\item{R_start_prior_sigma}{Prior (standard deviation) on the initial
reproduction number (intercept).}

\item{gp_sigma_prior_mu}{Prior (mean) on the magnitude of the Gaussian
process. Can be approximately interpreted as the marginal standard
deviation of changes in Rt. Higher values of sigma allow faster changes in
Rt.}

\item{gp_sigma_prior_sigma}{Prior (standard deviation) on the magnitude of
the Gaussian process.}

\item{gp_length_prior_mu}{Prior (mean) on the length scale of the Gaussian
process (in days). This influences the “persistence” of trends in Rt. A
higher length scale means that the Rt trend will change more slowly.
Choosing a length scale that is too short can lead to overfitting of the Rt
trajectory, while choosing a length scale that is too long can result in
unrealistically smooth Rt estimates.}

\item{gp_length_prior_sigma}{Prior (standard deviation) on the length scale
of the Gaussian process.}

\item{gp_length_max}{Maximum length scale of the Gaussian process. The normal
prior for the length scale is truncated between 0 and \code{gp_length_max}.}

\item{gp_matern_nu}{The smoothness parameter of the Matern kernel. The
default is 3/2, other possible choices are 5/2 (more smooth) and 1/2 (less
smooth). However, we recommend tuning smoothness primarily using the length
scale prior.}

\item{gp_boundary_factor}{The boundary factor used in the Gaussian process
approximation. The default (\code{gp_boundary_factor = 3}) is higher than the
minimum recommendation from Riutort-Mayol et al. to ensure accurate
real-time estimation. Lower values can lead to boundary effects close to
the present and start of the time series and are thus not recommended. When
modeling very long shedding delays, the boundary factor might need to be
further increased. Note that this will also automatically increase the
number of basis functions used and thereby slow down sampling.}

\item{gp_n_basis_factor}{Factor used to automatically determine \code{m}, the
number of basis functions in the Gaussian process approximation. Based on
recommendations from Riutort-Mayol et al., we let
\code{m = gp_n_basis_factor * c / (l / S)},
where \code{c} is the \code{gp_boundary_factor}, \code{l} is the length
scale of the GP (we use the 5\% quantile of \code{gp_length_prior} for a
conservative result), and \code{S} is the maximum absolute value of the input
space, which is \code{(n-1)/2} where \code{n} is the number of Rt time steps modeled.
This means that the number of basis function automatically adapts to the
length scale and number of observations in the model. For the default
settings (\code{gp_boundary_factor = 3} and \code{gp_n_basis_factor = 3.42}) and a
lower length scale of 3 weeks, this corresponds to approx. 0.25x the number
of Rt time points modeled. Increasing \code{gp_n_basis_factor} will make the
approximation more accurate by proportionally increasing the number of
basis functions, but can slow down sampling.}

\item{phi}{The strength of autocorrelation applied to the Gaussian process.
When \code{phi=1}, the Gaussian process effectively smoothes the first order
differences of Rt. This is desirable for real-time estimation, as it
encodes the minimally informed prior assumption that the current
transmission dynamics will continue unchanged. However the resulting
non-stationarity of the time series slows down MCMC sampling. In contrast,
for \code{phi<1}, the time series is stationary and reverts to \code{Rt=1} in the
long term. We here use a default of \code{phi=0.95}, which samples faster than
\code{phi=1} but still predicts continued transmission dynamics on short-term
scales. Decreasing \code{phi} will speed up sampling further but lead to faster
reversion to \code{Rt=1} towards the present (it is thus only recommended for
retrospective estimation, not real-time). Note that \code{phi=0} corresponds to
modeling a GP directly on Rt instead of the differences of Rt. This may
require adjusting the \code{gp_sigma} prior.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number Rt over
time as a smooth trend using a Gaussian process model.
}
\details{
The estimated Rt trajectory is primarily influenced by the priors
for the \emph{magnitude} and the \emph{length scale} of the Gaussian process. We
recommend adjusting the \emph{magnitude} when changes in Rt seem to be too slow
or too fast (often resulting in maximum Rt values that are too low or too
high). In contrast, we recommend adjusting the \emph{length scale} when the Rt
trajectory seems to have too much or too little resolution.

The Gaussian process is modeled using a Hilbert space approximation
as described in Riutort-Mayol et al. (2023). This allows for fast inference
without significant loss of accuracy. See the reference for more detail on
choosing an adequate boundary factor and basis function factor. The
\code{EpiSewer} implementation of the approximate Gaussian process is strongly
inspired by the implementation in the
\href{https://github.com/epiforecasts/EpiNow2/}{EpiNow2} package.

The priors of this component have the following functional form:
\itemize{
\item R_start (intercept):
\code{Normal}
\item gp_sigma (magnitude):
\verb{Truncated Normal}
\item gp_length (length scale):
\verb{Truncated Normal}
}
}
\references{
Riutort-Mayol, G., Bürkner, PC., Andersen, M.R. et al. Practical
Hilbert space approximate Bayesian Gaussian processes for probabilistic
programming. \emph{Stat Comput} 33, 17 (2023).
\url{https://doi.org/10.1007/s11222-022-10167-2}

Abbott S, Hellewell J, Thompson RN et al. Estimating the
time-varying reproduction number of SARS-CoV-2 using national and
subnational case counts. \emph{Wellcome Open Res} 5, 112 (2020).
\url{https://doi.org/10.12688/wellcomeopenres.16006.2}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_changepoint_splines}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_piecewise}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_smooth_derivative}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
