% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_gp}
\alias{R_estimate_gp}
\title{Estimate Rt using a Gaussian process}
\usage{
R_estimate_gp(
  R_start_prior_mu = 1,
  R_start_prior_sigma = 0.8,
  gp_sigma_prior_mu = 0.05,
  gp_sigma_prior_sigma = 0.01,
  gp_length_prior_mu = 28,
  gp_length_prior_sigma = 3.5,
  gp_matern_nu = 3/2,
  gp_padding_factor = 4,
  gp_phi = 0.9,
  link = "inv_softplus",
  R_max = 6,
  modeldata = modeldata_init()
)
}
\arguments{
\item{R_start_prior_mu}{Prior (mean) on the initial reproduction number
(intercept).}

\item{R_start_prior_sigma}{Prior (standard deviation) on the initial
reproduction number (intercept).}

\item{gp_sigma_prior_mu}{Prior (mean) on the magnitude of the Gaussian
process.}

\item{gp_sigma_prior_sigma}{Prior (standard deviation) on the magnitude of
the Gaussian process.}

\item{gp_length_prior_mu}{Prior (mean) on the length scale of the Gaussian
process.}

\item{gp_length_prior_sigma}{Prior (standard deviation) on the length scale
of the Gaussian process.}

\item{gp_matern_nu}{The Matern parameter nu of the Gaussian process. The
default is 3/2.}

\item{gp_padding_factor}{The domain of the Gaussian process must be padded to
attenuate boundary effects (resulting from a periodic kernel used for the
Fast Fourier Transform computation). The domain is padded both at the start
and end of the time series by \code{gp_padding_factor} times the length scale.
As expected length scale, we use \code{gp_length_prior_mu + gp_length_prior_sigma}.}

\item{gp_phi}{The damping factor of the Gaussian process. When \code{phi=1}, there
is no dampening and the Gaussian process is non-stationary (GP on first
order differences). This is desirable for real-time estimation, as it
encodes a prior assumption that the current transmission dynamics will
continue, however the non-stationarity slows down MCMC sampling. For
\code{phi<1}, the GP becomes stationary and reverts to the global mean in the
long term. We use a default of \code{phi=0.9}, which gives a stationary GP that
samples faster but still behaves similarly to a GP on first order
differences on short-term scales. Note that when \code{phi=0}, this corresponds
to modeling a GP directly on Rt instead of the differences of Rt. This may
require adjusting the \code{gp_sigma} prior.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number Rt over
time as a smooth trend using a Gaussian process model.
}
\details{
The priors of this component have the following functional form:
\itemize{
\item R_start (intercept):
\code{Normal}
\item gp_sigma (magnitude):
\verb{Truncated Normal}
\item gp_length (length scale):
\verb{Truncated Normal}
}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_changepoint_splines}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_piecewise}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_smooth_derivative}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
