% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_rw}
\alias{R_estimate_rw}
\title{Title}
\usage{
R_estimate_rw(
  intercept_prior_mu = 1,
  intercept_prior_sigma = 0.8,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.05,
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
