% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_rw}
\alias{R_estimate_rw}
\title{Estimate Rt via a random walk}
\usage{
R_estimate_rw(
  intercept_prior_mu = 1,
  intercept_prior_sigma = 0.8,
  sd_prior_mu = 0,
  sd_prior_sigma = 0.1,
  sd_changepoint_dist = 7 * 26,
  sd_changepoint_sd = 0.025,
  link = "inv_softplus",
  R_max = 6,
  differenced = FALSE,
  noncentered = TRUE,
  modeldata = modeldata_init()
)
}
\arguments{
\item{intercept_prior_mu}{Prior (mean) on the intercept of the random walk.}

\item{intercept_prior_sigma}{Prior (standard deviation) on the intercept of
the random walk.}

\item{sd_prior_mu}{Prior (mean) on the standard deviation of the random walk.}

\item{sd_prior_sigma}{Prior (standard deviation) on the standard deviation of
the random walk.}

\item{sd_changepoint_dist}{The variability of Rt can change over time, e.g.
during the height of an epidemic wave, countermeasures may lead to much
faster changes in Rt than observable at other times. This potential
variability is accounted for using change points placed at regular
intervals. The standard deviation of the random walk then evolves linearly
between the change points. The default change point distance is 26 weeks
(182 days). Short changepoint distances (e.g. 4 weeks or less) must be
chosen with care, as they can make the Rt time series too flexible. If set
to zero, no change points are modeled.}

\item{sd_changepoint_sd}{This parameter controls the variability of the
change points. When change points are modeled, EpiSewer will estimate a
baseline standard deviation (see \code{sd_prior_mu} and \code{sd_prior_sigma}), and
model change point values as independently distributed with mean equal to
this baseline and standard deviation \code{sd_changepoint_sd}.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{differenced}{If \code{FALSE} (default), the random walk is applied to the
absolute Rt time series. If \code{TRUE}, it is instead applied to the
differenced time series, i.e. now the trend is modeled as a random walk.}

\item{noncentered}{If \code{TRUE} (default), a non-centered parameterization is
used to model the innovations of the random walk (for better sampling
efficiency).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number over
time using a random walk.
}
\details{
The smoothness of Rt estimates is influenced by the prior on the
standard deviation of the random walk. It also influences the uncertainty
of Rt estimates towards the present / date of estimation, when limited data
signal is available. The prior on the intercept of the random walk should
reflect your expectation of Rt at the beginning of the time series. If
estimating from the start of an epidemic, you might want to use a prior
with mean > 1 for the intercept.

The priors of this component have the following functional form:
\itemize{
\item intercept of the random walk: \code{Normal}
\item standard deviation of the random walk: \verb{Truncated normal}
}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
