% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_infections.R
\name{R_estimate_smooth_derivative}
\alias{R_estimate_smooth_derivative}
\title{Estimate Rt with a smooth derivative}
\usage{
R_estimate_smooth_derivative(
  R_start_prior_mu = 1,
  R_start_prior_sigma = 0.8,
  smooth_changes_prior_mu = 0.02,
  smooth_changes_prior_sd = 0,
  spline_knot_distance = 14,
  sharp_changes = TRUE,
  sharp_changes_prior_shape = 0.5,
  sharp_changes_prior_rate = 1e-05,
  link = "inv_softplus",
  R_max = 6,
  modeldata = modeldata_init()
)
}
\arguments{
\item{R_start_prior_mu}{Prior (mean) on the initial reproduction number
(intercept).}

\item{R_start_prior_sigma}{Prior (standard deviation) on the initial
reproduction number (intercept).}

\item{smooth_changes_prior_mu}{Prior (mean) for the smoothness of the Rt
trend. Smaller values lead to smoother Rt trajectories by increasing the
difference penalty for the trend smoothing splines. This can be interpreted
as the standard deviation of a random walk over the spline coefficients. For
example, a value of 0.02 supports trend changes by up to approximately 0.04
per spline knot, which in turn corresponds to a change from a stable Rt=1 to
Rt=1.5 within two weeks.}

\item{smooth_changes_prior_sd}{Prior (sd) for the smoothness of the Rt trend.
By default, this is set to 0, which means that the smoothness is fixed to
\code{smooth_changes_prior_mu} and not estimated from the data.}

\item{spline_knot_distance}{Distance (in days) between spline knots for the
penalized smoothing splines. Shorter distances increase flexibility of the
Rt trajectory but also the daily Rt uncertainty.}

\item{sharp_changes}{Should occasional fast changes in Rt be supported? This
option multiplies each trend spline coefficient by a factor of 1+alpha_i,
where each alpha_i is sampled from a sparse distribution. This allows
individual trend spline coefficients to be extremely high (positive trend)
or low (negative trend), leading to fast short-term changes in Rt.}

\item{sharp_changes_prior_shape}{Exponential-Gamma (EG) prior (shape) for the
strength of sharp changes in Rt. The exponential-Gamma prior is sparse, i.e.
it has a strong peak towards zero (no deviation from penalized splines) and
a long tail (occassional sharp changes). Smaller shape parameters will lead
to more sparseness, i.e. a longer tail. Note that when adjusting the shape,
you will likely also have to adjust the rate.}

\item{sharp_changes_prior_rate}{Exponential-Gamma (EG) prior (rate) for the
strength of sharp changes in the Rt trend. See \code{sharp_changes_prior_shape}
above for an explanation. Larger rates will lead to more variability.}

\item{link}{Link function. Currently supported are \code{inv_softplus} (default)
and \code{scaled_logit}. Both of these links are configured to behave
approximately like the identity function around R=1, but become increasingly
non-linear below (and in the case of \code{scaled_logit} also above) R=1.}

\item{R_max}{If \code{link=scaled_logit} is used, a maximum reproduction number
must be assumed. This should be higher than any realistic R value for the
modeled pathogen. Default is 6.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{.init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option estimates the effective reproduction number Rt over
time as a smooth trend. It uses penalized smoothing splines placed on the
first-order differences of Rt. This leads to mostly gradual changes.
Occasional sharp changes are supported through an additional sparse prior.
}
\details{
The priors of this component have the following functional form:
\itemize{
\item R_start (intercept):
\code{Normal}
\item smooth_changes (standard deviation of the random walk over trend spline coefficients):
\code{Truncated-normal}
\item sharp_changes (additional multiplier to spline coefficients):
\code{Exponential-Gamma}
}
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_approx}()},
\code{\link{R_estimate_changepoint_splines}()},
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_piecewise}()},
\code{\link{R_estimate_rw}()},
\code{\link{R_estimate_splines}()}
}
\concept{{Rt models}}
