% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_splines}
\alias{R_estimate_splines}
\title{Title}
\usage{
R_estimate_splines(
  standata = standata_init(),
  knot_distance = 1,
  spline_degree = 3,
  bs_coeff_ar_start_prior = stan_prior("bs_coeff_ar_start", "normal", mu = 0, sigma =
    0.5),
  bs_coeff_ar_sd_prior = stan_prior("bs_coeff_ar_sd", "half-normal", mu = 0, sigma =
    0.2)
)
}
\arguments{
\item{bs_coeff_ar_sd_prior}{}
}
\value{

}
\description{
Title
}
