% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_splines}
\alias{R_estimate_splines}
\title{Smooth Rt estimates via smoothing splines}
\usage{
R_estimate_splines(
  knot_distance = 1,
  spline_degree = 3,
  coef_intercept_prior_mu = 0,
  coef_intercept_prior_sigma = 0.5,
  coef_sd_prior_mu = 0,
  coef_sd_prior_sigma = 0.2,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}

\item{bs_coeff_ar_sd_prior}{}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Smooth Rt estimates via smoothing splines
}
\seealso{
Other {Rt models}: 
\code{\link{R_estimate_ets}()},
\code{\link{R_estimate_rw}()}
}
\concept{{Rt models}}
