% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{R_estimate_splines}
\alias{R_estimate_splines}
\title{Title}
\usage{
R_estimate_splines(
  knot_distance = 1,
  spline_degree = 3,
  coef_intercept_prior_mu = 0,
  coef_intercept_prior_sigma = 0.5,
  coef_sd_prior_mu = 0,
  coef_sd_prior_sigma = 0.2,
  modeldata = modeldata_init()
)
}
\arguments{
\item{bs_coeff_ar_sd_prior}{}
}
\value{

}
\description{
Title
}
