% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{residence_dist_assume}
\alias{residence_dist_assume}
\title{Title}
\usage{
residence_dist_assume(residence_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
