% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sewage.R
\name{residence_dist_assume}
\alias{residence_dist_assume}
\title{Assume a sewer residence time distribution}
\usage{
residence_dist_assume(residence_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{residence_dist}{A numeric vector representing a discrete residence time
distribution, with elements describing the share of load that takes 0 days,
1 day, 2 days, and so on to arrive at the sampling site.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option assumes a fixed residence time distribution for
pathogen particles. By default, \code{EpiSewer} assumes that particles arrive at
the sampling site within the day of shedding. However, for larger sewage
systems, particles may travel longer than a day depending on where and
when they were shed into the wastewater.
}
\examples{
# Particles arrive within the same day
residence_dist_assume(residence_dist = c(1))

# Particles always arrive after one day
residence_dist_assume(residence_dist = c(0, 1))

1/4 of particles only arrives after one day
residence_dist_assume(residence_dist = c(0.75, 0.25))
}
