% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{sample_effects_estimate_matrix}
\alias{sample_effects_estimate_matrix}
\title{Estimate sample effects using a design matrix}
\usage{
sample_effects_estimate_matrix(
  design_matrix,
  effect_prior_mu = 0,
  effect_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{design_matrix}{}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Estimate sample effects using a design matrix
}
\seealso{
Other {sample effect models}: 
\code{\link{sample_effects_estimate_weekday}()},
\code{\link{sample_effects_none}()}
}
\concept{{sample effect models}}
