% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{sample_effects_estimate_matrix}
\alias{sample_effects_estimate_matrix}
\title{Title}
\usage{
sample_effects_estimate_matrix(
  design_matrix,
  effect_prior_mu = 0,
  effect_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{design_matrix}{}
}
\value{

}
\description{
Title
}
