% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{sample_effects_estimate_weekday}
\alias{sample_effects_estimate_weekday}
\title{Title}
\usage{
sample_effects_estimate_weekday(
  standata = standata_init(),
  eta_prior = stan_prior("eta", "normal", mu = 0, sigma = 1)
)
}
\arguments{
\item{eta_prior}{}
}
\value{

}
\description{
Title
}
