% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{sample_effects_estimate_weekday}
\alias{sample_effects_estimate_weekday}
\title{Estimate weekday sample effects}
\usage{
sample_effects_estimate_weekday(
  effect_prior_mu = 0,
  effect_prior_sigma = 1,
  modeldata = modeldata_init()
)
}
\arguments{
\item{effect_prior_mu}{Prior (mean) on the regression coefficients.}

\item{effect_prior_sigma}{Prior (standard deviation) on the regression
coefficients.}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
This option uses a linear regression model to estimate sample
weekday effects on the concentration. Concentrations can be influenced by
the time between sampling and shipping to the lab (age-of-sample effect),
and if shipment follows a weekly batch scheme, the sampling weekday is a
good proxy for the age at shipment.
}
\details{
Effects are estimated for weekdays Monday - Saturday, with Sunday as
the baseline. \code{EpiSewer} will fit a fixed-effects linear model, random
effects are currently not supported.

The priors of this component have the following functional form:
\itemize{
\item regression coefficients: \code{Normal}
}
}
\seealso{
Other {sample effect models}: 
\code{\link{sample_effects_estimate_matrix}()},
\code{\link{sample_effects_none}()}
}
\concept{{sample effect models}}
