% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{seeding_estimate}
\alias{seeding_estimate}
\title{Title}
\usage{
seeding_estimate(
  iota_log_ar_start_prior = NULL,
  iota_log_ar_sd_prior = set_prior("iota_log_ar_sd", "half-normal", mu = 0.05, sigma =
    0.025),
  modeldata = modeldata_init()
)
}
\arguments{
\item{iota_log_ar_sd_prior}{}
}
\value{

}
\description{
Title
}
