% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{seeding_estimate}
\alias{seeding_estimate}
\title{Title}
\usage{
seeding_estimate(
  intercept_prior_mu = NULL,
  intercept_prior_sigma = NULL,
  sd_prior_mu = 0.05,
  sd_prior_sigma = 0.025,
  modeldata = modeldata_init()
)
}
\arguments{
\item{modeldata}{}
}
\value{

}
\description{
Title
}
