% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{set_fit_opts}
\alias{set_fit_opts}
\title{Configure the model fitting}
\usage{
set_fit_opts(
  sampler = sampler_stan_mcmc(),
  fitted = TRUE,
  model = model_stan_opts()
)
}
\arguments{
\item{sampler}{Which type of sampler should be used for model fitting?
Currently, only \code{\link[=sampler_stan_mcmc]{sampler_stan_mcmc()}} is supported.}

\item{fitted}{If \code{TRUE} (default), the fitted model object is also returned,
not only summaries of the model fitting. Note that this makes the results
object substantially larger.}

\item{model}{Details about the model file to be used, see
\code{\link[=model_stan_opts]{model_stan_opts()}}. This also allows to use model files customized by the
user.}
}
\value{
A \code{list} with the model fitting options.
}
\description{
Sets model fitting options in \code{EpiSewer}.
}
