% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stan.R
\name{set_fit_opts}
\alias{set_fit_opts}
\title{Configure the model fitting}
\usage{
set_fit_opts(sampler = sampler_stan_mcmc(), model = model_stan_opts())
}
\arguments{
\item{sampler}{Which type of sampler should be used for model fitting?
Currently, only \code{\link[=sampler_stan_mcmc]{sampler_stan_mcmc()}} is supported.}

\item{model}{Details about the model file to be used, see
\code{\link[=model_stan_opts]{model_stan_opts()}}. This also makes it possible for users to supply
customized models.}

\item{results}{Details about the results to be returned, see
\code{\link[=results_opts]{results_opts()}}. For example, what credible intervals to use for
summarizing posterior distributions or whether to return the fitted model
object.}
}
\value{
A \code{list} with the model fitting options.
}
\description{
Sets model fitting options in \code{EpiSewer}.
}
\examples{
ww_data <- ww_data_SARS_CoV_2_Zurich
ww_assumptions <- ww_assumptions_SARS_CoV_2_Zurich

ww_fit_opts <- set_fit_opts(
  sampler = sampler_stan_mcmc(
    chains = 2,
    parallel_chains = 2,
    iter_warmup = 400,
    iter_sampling = 400,
    seed = 42 # ensures reproducibility
  )
)

ww_results_opts <- set_results_opts(
  fitted = TRUE, # return full model fit
  summary_intervals = c(0.5, 0.8, 0.95),
  samples_ndraws = 100
)

ww_result <- EpiSewer(
  data = ww_data,
  assumptions = ww_assumptions,
  fit_opts = ww_fit_opts,
  results_opts = ww_results_opts
)
}
