% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{sewer_assumptions}
\alias{sewer_assumptions}
\title{Title}
\usage{
sewer_assumptions(
  generation_dist = NULL,
  incubation_dist = NULL,
  shedding_dist = NULL,
  load_per_case = NULL,
  ...
)
}
\arguments{
\item{load_per_case}{}
}
\value{

}
\description{
Title
}
