% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiSewer.R
\name{sewer_assumptions}
\alias{sewer_assumptions}
\title{Specify modeling assumptions}
\usage{
sewer_assumptions(
  generation_dist = NULL,
  incubation_dist = NULL,
  shedding_dist = NULL,
  min_cases = 10,
  load_per_case = NULL,
  residence_dist = c(1),
  ...
)
}
\arguments{
\item{generation_dist}{Generation time distribution. The intrinsic
distribution of the time between infection of a primary case and infection
of its secondary cases. Will be automatically passed to
\code{\link[=generation_dist_assume]{generation_dist_assume()}}.}

\item{incubation_dist}{Incubation period distribution. \code{EpiSewer} uses this
as a proxy for the time between infection and the start of shedding, as
shedding load distributions in the literature are often given from symptom
onset onwards. If the supplied shedding load distribution instead starts
with the time of infection, use \code{incubation_dist=c(1)} (i.e. no lag). Will
be automatically passed to \code{\link[=incubation_dist_assume]{incubation_dist_assume()}}.}

\item{shedding_dist}{Shedding load distribution. Describes how the total load
shed by an individual is distributed over time (and therefore sums to 1).
Will be automatically passed to \code{\link[=shedding_dist_assume]{shedding_dist_assume()}}.}

\item{min_cases}{The minimum incidence (new cases per day) during the modeled
time period (default is 10). This assumption is used to calibrate the
\code{load_per_case} factor (a scaling factor that describes how many pathogen
particles are shed by the average infected individual overall and how much
of this is detectable at the sampling site). The \code{load_per_case} depends
both on biological factors as well as on the specific sewage system and
laboratory quantification. If \code{min_cases} is specified, \code{load_per_case} is
chosen such that the estimated time series of infections does not go
significantly below \code{min_cases}. Will be automatically passed to
\code{\link[=load_per_case_calibrate]{load_per_case_calibrate()}}.}

\item{load_per_case}{This argument allows to directly specify the average
total load per case, instead of calibrating it to case data. Important: To
actually use this assumption, the model option \code{\link[=load_per_case_assume]{load_per_case_assume()}}
must be used. See \code{\link[=suggest_load_per_case]{suggest_load_per_case()}} if you want to do a manual
calibration to case data.}

\item{residence_dist}{Sewer residence time distribution for pathogen
particles. By default, \code{EpiSewer} assumes that particles arrive at the
sampling site within the day of shedding. However, for larger sewage
systems, particles may travel longer than a day depending on where and when
they were shed into the wastewater. Will be automatically passed to
\code{\link[=residence_dist_assume]{residence_dist_assume()}}.}

\item{...}{Further assumptions to be passed to \code{\link[=EpiSewer]{EpiSewer()}}.}
}
\value{
A \code{list} with all assumptions supplied. Can be passed to the
\code{assumptions} argument in \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Specify model assumptions such as the generation time
distribution or shedding load distribution. This is a convenience function
to collect all assumptions in one object.
}
\details{
Note that if case data is supplied via \code{sewer_data}, the \code{min_cases}
assumption is overwritten and the supplied case data is used for a more
accurate calibration instead.
}
