% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{shedding_dist_assume}
\alias{shedding_dist_assume}
\title{Title}
\usage{
shedding_dist_assume(standata = standata_init(), shedding_dist)
}
\arguments{
\item{shedding_dist}{}
}
\value{

}
\description{
Title
}
