% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_define.R
\name{shedding_dist_assume}
\alias{shedding_dist_assume}
\title{Aasume a shedding load profile / distribution}
\usage{
shedding_dist_assume(shedding_dist = NULL, modeldata = modeldata_init())
}
\arguments{
\item{shedding_dist}{}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
model helpers.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} model helpers to add
further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{meta_info}), and checks to be
performed before model fitting (\code{checks}).
}
\description{
Aasume a shedding load profile / distribution
}
