% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{shedding_dist_assume}
\alias{shedding_dist_assume}
\title{Assume a shedding load distribution}
\usage{
shedding_dist_assume(
  shedding_dist = NULL,
  shedding_reference = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{shedding_dist}{A numeric vector representing a discrete shedding load
distribution, with elements describing the share of load shed 0 days, 1
day, 2 days, and so on after the start of shedding.}

\item{shedding_reference}{Is the shedding load distribution relative to the
day of \code{"infection"} or the day of \code{"symptom_onset"}? This is important because
shedding load distributions provided in the literature are sometimes by
days since infection and sometimes by days since symptom onset. If
\code{shedding_reference="symptom_onset"}, EpiSewer also needs information about
the incubation period distribution (see \code{\link[=incubation_dist_assume]{incubation_dist_assume()}}).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option assumes a fixed shedding load distribution for the
shedding model in \code{EpiSewer}.
}
\seealso{
Helpers to discretize continuous probability distributions:
\code{\link[=get_discrete_gamma]{get_discrete_gamma()}}, \code{\link[=get_discrete_lognormal]{get_discrete_lognormal()}}
}
