% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shedding.R
\name{shedding_dist_estimate}
\alias{shedding_dist_estimate}
\title{Estimate an uncertain shedding load distribution}
\usage{
shedding_dist_estimate(
  shedding_dist_mean_prior_mean = NULL,
  shedding_dist_mean_prior_sd = NULL,
  shedding_dist_cv_prior_mean = NULL,
  shedding_dist_cv_prior_sd = NULL,
  shedding_dist_type = "gamma",
  shedding_reference = NULL,
  modeldata = modeldata_init()
)
}
\arguments{
\item{shedding_dist_mean_prior_mean}{Prior (mean) for the mean of the
shedding load distribution.}

\item{shedding_dist_mean_prior_sd}{Prior (standard deviation) for the mean of
the shedding load distribution.}

\item{shedding_dist_cv_prior_mean}{Prior (mean) for the coefficient of
variation (i.e. standard deviation relative to the mean) of the shedding
load distribution.}

\item{shedding_dist_cv_prior_sd}{Prior (standard deviation) for the
coefficient of variation of the shedding load distribution.}

\item{shedding_dist_type}{The parametric distribution that should be modeled.
Supported are "gamma", "exponential", and "lognormal".}

\item{shedding_reference}{Is the shedding load distribution relative to the
day of \code{"infection"} or the day of \code{"symptom_onset"}? This is important
because shedding load distributions provided in the literature are
sometimes by days since infection and sometimes by days since symptom
onset. If \code{shedding_reference="symptom_onset"}, EpiSewer also needs
information about the incubation period distribution (see
\code{\link[=incubation_dist_assume]{incubation_dist_assume()}}).}

\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
A \code{modeldata} object containing data and specifications of the model
to be fitted. Can be passed on to other \code{EpiSewer} modeling functions to
add further data and model specifications.

The \code{modeldata} object also includes information about parameter
initialization (\code{init}), meta data (\code{.metainfo}), and checks to be
performed before model fitting (\code{.checks}).
}
\description{
This option models a parametric shedding load distribution. It
is useful for representing uncertainty about the mean and variation of the
shedding profile.
}
