% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess.R
\name{suggest_load_per_case}
\alias{suggest_load_per_case}
\title{Title}
\usage{
suggest_load_per_case(
  measurements,
  flows,
  cases,
  ascertainment_prop = 1,
  measurement_shift = seq(-7, 7),
  shift_weights = 1/(abs(measurement_shift) + 1),
  date_col = "date",
  measurement_col = "concentration",
  flow_col = "flow",
  case_col = "cases"
)
}
\arguments{
\item{flow_col}{}
}
\value{

}
\description{
Title
}
