% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewer_summary.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Summarize parameters of interest}
\usage{
summarize_fit(fit, data, .metainfo, intervals = c(0.5, 0.95), ndraws = 50)
}
\arguments{
\item{fit}{The fitted \code{EpiSewer} model.}

\item{data}{The model data that the \code{EpiSewer} model was fitted with.}

\item{.metainfo}{Meta information about the model data.}

\item{intervals}{The credible intervals (CrIs) that should be calculated. By
default, these are the 50\% and 95\% CrIs.}

\item{ndraws}{Number of exemplary posterior samples that should be extracted.
(The summaries always use all draws.)}
}
\value{
A \code{list} with the following summaries (each in a \code{data.frame}):
\itemize{
\item R (posterior summary)
\item R_samples (exemplary samples)
\item expected_infections (posterior summary)
\item infections (posterior summary)
\item infections_samples (exemplary samples)
\item expected_load (posterior summary)
\item expected_concentration (posterior summary)
\item concentration (posterior summary)
\item sample_effects (posterior summary)
}
}
\description{
This function summarizes important parameters of interest from a
fitted \code{EpiSewer} model.
}
\details{
The summaries for infections and R include a column \code{seeding}, which
indicates whether the corresponding date was still in the seeding phase or
not. The seeding phase is here defined as 2xG days long, where G is the
maximum generation time. This is in contrast to the model, where it is only
G days long. The reason for this decision is that after G days, new
infections are still based on the seeded infections and not strongly
informed by the data. By using the extended seeding criterion, we ensure
that only sufficiently data-driven estimates are shown in the result plots.
}
