% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy.R
\name{tbp}
\alias{tbp}
\title{To be provided}
\usage{
tbp(
  f_name,
  f_expr,
  required_data = c(),
  required_assumptions = c(),
  modeldata = NULL,
  calling_env = rlang::caller_env()
)
}
\description{
To be provided
}
