% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modeldata.R
\name{template_model_helpers}
\alias{template_model_helpers}
\title{Template function}
\usage{
template_model_helpers(modeldata)
}
\arguments{
\item{modeldata}{A \code{modeldata} object to which the above model specifications
should be added. Default is an empty model given by \code{\link[=modeldata_init]{modeldata_init()}}. Can
also be an already partly specified model returned by other \code{EpiSewer}
modeling functions.}
}
\value{
Nothing
}
\description{
This function does not actually do anything. It only serves as a
template for the documentation of other functions using inheritance of
params.
}
