% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_assumptions_SARS_CoV_2_Zurich}
\alias{ww_assumptions_SARS_CoV_2_Zurich}
\title{Example sewer_assumptions object for SARS-CoV-2 in Zurich}
\format{
A \code{sewer_assumptions} object as returned by \code{\link[=sewer_assumptions]{sewer_assumptions()}}.
}
\description{
Epidemiological assumptions for SARS-CoV-2, including
generation time, incubation period, and shedding load distributions
derived from the literature.
}
\seealso{
\code{\link{ww_data_SARS_CoV_2_Zurich}}, \code{\link{ww_result_SARS_CoV_2_Zurich}}
}
\keyword{datasets}
