% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_data_influenza_Zurich}
\alias{ww_data_influenza_Zurich}
\title{Example sewer_data object for Influenza A in Zurich}
\format{
A \code{sewer_data} object as returned by \code{\link[=sewer_data]{sewer_data()}}.
}
\description{
Processed wastewater data for Influenza A from the WWTP
Werdhölzli in Zürich (September–December 2022), ready for use with
\code{\link[=EpiSewer]{EpiSewer()}}.
}
\seealso{
\code{\link{Influenza_A_Zurich}}, \code{\link{ww_assumptions_influenza_Zurich}}
}
\keyword{datasets}
\keyword{internal}
