% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_result_SARS_CoV_2_Zurich}
\alias{ww_result_SARS_CoV_2_Zurich}
\title{Example EpiSewerJobResult for SARS-CoV-2 in Zurich}
\format{
An \code{EpiSewerJobResult} object as returned by \code{\link[=EpiSewer]{EpiSewer()}}.
}
\description{
Result of running \code{\link[=EpiSewer]{EpiSewer()}} on
\code{\link{ww_data_SARS_CoV_2_Zurich}} with
\code{\link{ww_assumptions_SARS_CoV_2_Zurich}}. Can be used to explore
outputs and plotting functions without fitting a model.
}
\seealso{
\code{\link{ww_data_SARS_CoV_2_Zurich}}, \code{\link{ww_assumptions_SARS_CoV_2_Zurich}}
}
\keyword{datasets}
